Add WorkflowError class and implement GitHub thread response publishing
- Introduced a new `WorkflowError` class for handling workflow-related errors with detailed attributes.
- Created a `publisher.py` module to manage the publishing of GitHub thread responses.
- Implemented logic to handle various scenarios during the publishing process, including missing responses and thread IDs.
- Added telemetry configuration for publishing sessions and structured logging of publish attempts and errors.
- Enhanced error handling to raise `WorkflowError` with appropriate status and reason codes for better debugging.
