+++ src/gh_address_cr/cli.py
@@ -15,15 +15,16 @@ from datetime import datetime, timezone
 from pathlib import Path
 from uuid import uuid4
 
-from gh_address_cr import (
-    MAX_PARALLEL_CLAIMS,
-    PROTOCOL_VERSION,
-    SUPPORTED_PROTOCOL_VERSIONS,
-    SUPPORTED_SKILL_CONTRACT_VERSIONS,
-    __version__,
+from gh_address_cr import __version__
+from gh_address_cr.commands.common import (
+    emit_scope_resolution_error as _emit_scope_resolution_error,
+    maybe_prepend_implicit_scope as _maybe_prepend_implicit_scope,
+    root_passthrough_args as _root_passthrough_args,
 )
+from gh_address_cr.commands.agent import handle_agent_command
+from gh_address_cr.commands.final_gate import handle_final_gate
+from gh_address_cr.commands.telemetry import handle_telemetry_command
 from gh_address_cr.core import gate as core_gate
-from gh_address_cr.core import telemetry as core_telemetry
 from gh_address_cr.core.github_thread_state import (
     is_claimable_github_thread,
     is_resolved_github_thread,
@@ -33,7 +34,6 @@ from gh_address_cr.core.handoff import (
     ensure_handoff_state as _ensure_handoff_state,
     record_producer_result,
 )
-from gh_address_cr.core import paths as core_paths
 from gh_address_cr.core import session as session_store
 from gh_address_cr.core.severity import apply_severity_evidence, severity_evidence
 from gh_address_cr.core import workflow
@@ -66,38 +66,6 @@ PUBLIC_COMMANDS = {
     "telemetry",
 }
 PR_SCOPED_IMPLICIT_COMMANDS = {"address", "review", "threads", "final-gate"}
-IMPLICIT_SCOPE_VALUE_OPTIONS = {
-    "--agent-id",
-    "--audit-id",
-    "--commit",
-    "--concern-label",
-    "--file",
-    "--files",
-    "--format",
-    "--handoff-sha256",
-    "--head",
-    "--homogeneous-reason",
-    "--input",
-    "--item-id",
-    "--max-iterations",
-    "--name",
-    "--note",
-    "--now",
-    "--repo",
-    "--review-priority",
-    "--role",
-    "--scan-id",
-    "--severity",
-    "--severity-note",
-    "--severity-override-note",
-    "--snapshot",
-    "--source",
-    "--test-command",
-    "--test-result",
-    "--validation",
-    "--validation-cmd",
-    "--why",
-}
 UNSUPPORTED_LEGACY_COMMANDS = {
     "audit-report",
     "batch-resolve",
@@ -124,10 +92,6 @@ INPUT_REQUIRED_COMMANDS = {"findings"}
 WAITING_FOR_EXTERNAL_REVIEW_EXIT = 6
 PR_IO_PREFLIGHT_EXIT = 5
 PR_URL_RE = re.compile(r"^https?://github\.com/(?P<owner>[^/]+)/(?P<repo>[^/]+)/pull/(?P<pr_number>\d+)(?:[/?#].*)?$")
-HOST_TELEMETRY_INPUT_ENV = "GH_ADDRESS_CR_HOST_TELEMETRY_INPUT"
-HOST_TELEMETRY_SOURCE_ENV = "GH_ADDRESS_CR_HOST_TELEMETRY_SOURCE"
-HOST_TELEMETRY_FORMAT_ENV = "GH_ADDRESS_CR_HOST_TELEMETRY_FORMAT"
-
 
 def _legacy_module(name: str):
     raise RuntimeError(f"Legacy module imports are not supported by the native CLI path: {name}")
@@ -1682,1251 +1646,6 @@ def _emit_native_summary(summary: dict, *, human: bool) -> None:
     sys.stdout.write(json.dumps(summary, indent=2, sort_keys=True) + "\n")
 
 
-def build_agent_manifest() -> dict:
-    return {
-        "status": "MANIFEST_READY",
-        "schema_version": PROTOCOL_VERSION,
-        "runtime_package": "gh-address-cr",
-        "runtime_version": __version__,
-        "agent_id": "gh-address-cr-runtime",
-        "protocol_versions": list(SUPPORTED_PROTOCOL_VERSIONS),
-        "supported_protocol_versions": list(SUPPORTED_PROTOCOL_VERSIONS),
-        "supported_skill_contract_versions": list(SUPPORTED_SKILL_CONTRACT_VERSIONS),
-        "roles": [
-            "coordinator",
-            "review_producer",
-            "triage",
-            "fixer",
-            "verifier",
-            "publisher",
-            "gatekeeper",
-        ],
-        "actions": [
-            "review",
-            "produce_findings",
-            "triage",
-            "classify",
-            "fix",
-            "trivial_fix",
-            "fix_all",
-            "evidence",
-            "clarify",
-            "defer",
-            "reject",
-            "verify",
-            "publish",
-            "gate",
-            "resolve_stale",
-        ],
-        "input_formats": [
-            "action_request.v1",
-            "finding.v1",
-            "github_thread.v1",
-            "evidence_profile.v1",
-            "workflow_decision.v1",
-        ],
-        "output_formats": [
-            "action_response.v1",
-            "batch_action_response.v1",
-            "evidence_record.v1",
-            "evidence_profile.v1",
-            "gate_report.v1",
-            "workflow_decision.v1",
-        ],
-        "constraints": {
-            "max_parallel_claims": MAX_PARALLEL_CLAIMS,
-        },
-        "public_commands": sorted(PUBLIC_COMMANDS),
-    }
-
-
-def handle_agent_command(args: argparse.Namespace) -> int:
-    if args.repo in {None, "-h", "--help"}:
-        sys.stdout.write(
-            "usage: gh-address-cr agent {manifest,classify,next,submit,submit-batch,fix,trivial-fix,fix-all,resolve-stale,evidence,publish,leases,reclaim,orchestrate} ...\n\n"
-            "Agent protocol utilities.\n"
-        )
-        return 0
-    if args.repo == "manifest" and not args.pr_number and not args.args:
-        sys.stdout.write(json.dumps(build_agent_manifest(), indent=2, sort_keys=True) + "\n")
-        return 0
-    if args.repo == "classify":
-        return handle_agent_classify(args.pr_number, args.args)
-    if args.repo == "next":
-        return handle_agent_next(args.pr_number, args.args)
-    if args.repo == "submit":
-        return handle_agent_submit(args.pr_number, args.args)
-    if args.repo == "submit-batch":
-        return handle_agent_submit_batch(args.pr_number, args.args)
-    if args.repo == "fix":
-        return handle_agent_fix(args.pr_number, args.args)
-    if args.repo == "trivial-fix":
-        return handle_agent_trivial_fix(args.pr_number, args.args)
-    if args.repo == "fix-all":
-        return handle_agent_fix_all(args.pr_number, args.args)
-    if args.repo == "resolve-stale":
-        return handle_agent_resolve_stale(args.pr_number, args.args)
-    if args.repo == "evidence":
-        return handle_agent_evidence(args.pr_number, args.args)
-    if args.repo == "publish":
-        return handle_agent_publish(args.pr_number, args.args)
-    if args.repo == "leases":
-        return handle_agent_leases(args.pr_number, args.args)
-    if args.repo == "reclaim":
-        return handle_agent_reclaim(args.pr_number, args.args)
-    if args.repo == "orchestrate":
-        return handle_agent_orchestrate(args.pr_number, args.args)
-    print(
-        "Unknown agent command. Supported commands: manifest, classify, next, submit, submit-batch, fix, trivial-fix, fix-all, resolve-stale, evidence, publish, leases, reclaim, orchestrate.",
-        file=sys.stderr,
-    )
-    return 2
-
-
-def handle_agent_classify(repo: str | None, passthrough: list[str]) -> int:
-    parser = argparse.ArgumentParser(prog="gh-address-cr agent classify")
-    parser.add_argument("repo")
-    parser.add_argument("pr_number")
-    parser.add_argument("item_id")
-    parser.add_argument("--classification", required=True, choices=["fix", "clarify", "defer", "reject"])
-    parser.add_argument("--agent-id", default="agent")
-    parser.add_argument("--note", required=True)
-    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
-    try:
-        payload = workflow.record_classification(
-            parsed.repo,
-            parsed.pr_number,
-            item_id=parsed.item_id,
-            classification=parsed.classification,
-            agent_id=parsed.agent_id,
-            note=parsed.note,
-        )
-    except workflow.WorkflowError as exc:
-        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
-    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-    return 0
-
-
-def handle_agent_next(repo: str | None, passthrough: list[str]) -> int:
-    parser = argparse.ArgumentParser(prog="gh-address-cr agent next")
-    parser.add_argument("repo")
-    parser.add_argument("pr_number")
-    parser.add_argument("--role", required=True)
-    parser.add_argument("--agent-id", default="agent")
-    parser.add_argument("--item-id")
-    parser.add_argument("--now")
-    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
-    try:
-        now_dt = None
-        if parsed.now:
-            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
-        payload = workflow.issue_action_request(
-            parsed.repo,
-            parsed.pr_number,
-            role=parsed.role,
-            agent_id=parsed.agent_id,
-            item_id=parsed.item_id,
-            now=now_dt,
-        )
-    except workflow.WorkflowError as exc:
-        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
-    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-    return 0
-
-
-def handle_agent_submit(repo: str | None, passthrough: list[str]) -> int:
-    parser = argparse.ArgumentParser(prog="gh-address-cr agent submit")
-    parser.add_argument("repo")
-    parser.add_argument("pr_number")
-    parser.add_argument("--input", required=True)
-    parser.add_argument("--publish", action="store_true", help="Publish accepted GitHub-thread fix evidence immediately.")
-    parser.add_argument("--now")
-    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
-    try:
-        now_dt = None
-        if parsed.now:
-            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
-        payload = workflow.submit_action_response(
-            parsed.repo,
-            parsed.pr_number,
-            response_path=parsed.input,
-            now=now_dt,
-            publish=parsed.publish,
-        )
-    except workflow.WorkflowError as exc:
-        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
-    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-    return 0
-
-
-def handle_agent_submit_batch(repo: str | None, passthrough: list[str]) -> int:
-    parser = argparse.ArgumentParser(
-        prog="gh-address-cr agent submit-batch",
-        description="Submit a BatchActionResponse with common fix evidence for multiple GitHub review threads.",
-    )
-    parser.add_argument("repo")
-    parser.add_argument("pr_number")
-    parser.add_argument("--input", required=True, help="Path to a BatchActionResponse JSON file.")
-    parser.add_argument("--now")
-    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
-    try:
-        now_dt = None
-        if parsed.now:
-            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
-        payload = workflow.submit_batch_action_response(
-            parsed.repo,
-            parsed.pr_number,
-            batch_path=parsed.input,
-            now=now_dt,
-        )
-    except workflow.WorkflowError as exc:
-        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
-    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-    return 0
-
-
-def _parse_agent_files(files: str | None, extra_files: list[str] | None = None) -> list[str]:
-    values: list[str] = []
-    if files:
-        values.extend(part.strip() for part in files.split(",") if part.strip())
-    for item in extra_files or []:
-        values.extend(part.strip() for part in item.split(",") if part.strip())
-    return values
-
-
-def _parse_agent_validation(values: list[str] | None) -> list[dict[str, str]]:
-    commands: list[dict[str, str]] = []
-    for raw in values or []:
-        command, result = _split_agent_validation_record(raw.strip())
-        if command and result:
-            commands.append({"command": command, "result": result})
-    return commands
-
-
-def _split_agent_validation_record(raw: str) -> tuple[str, str]:
-    command, separator, result = raw.rpartition("=")
-    if not separator or not _looks_like_agent_validation_result(result):
-        return raw.strip(), "passed"
-    return command.strip(), result.strip()
-
-
-def _looks_like_agent_validation_result(value: str) -> bool:
-    normalized = value.strip().lower()
-    if not normalized or any(char.isspace() for char in normalized):
-        return False
-    return normalized in {"pass", "passed", "success", "succeeded", "ok", "fail", "failed", "error", "skipped"}
-
-
-def _changed_files_for_commit(
-    commit_hash: str,
-    *,
-    rejected_status: str = "FAST_FIX_ALL_REJECTED",
-    command_name: str = "agent fix-all",
-) -> list[str]:
-    commit = commit_hash.strip()
-    if not commit:
-        return []
-    if commit.startswith("-"):
-        raise workflow.WorkflowError(
-            status=rejected_status,
-            reason_code="INVALID_COMMIT_HASH",
-            waiting_on="git_commit",
-            exit_code=2,
-            message=f"{command_name} requires a commit-ish that does not start with '-'.",
-        )
-    commands = [
-        ["git", "diff-tree", "--no-commit-id", "--name-only", "-r", commit],
-        ["git", "show", "--format=", "--name-only", commit],
-    ]
-    last_error = ""
-    for command in commands:
-        result = subprocess.run(command, text=True, capture_output=True)
-        if result.returncode != 0:
-            last_error = result.stderr.strip() or result.stdout.strip()
-            continue
-        files = [line.strip() for line in result.stdout.splitlines() if line.strip()]
-        if files:
-            return files
-    raise workflow.WorkflowError(
-        status=rejected_status,
-        reason_code="COMMIT_FILES_UNAVAILABLE",
-        waiting_on="git_commit",
-        exit_code=2,
-        message=last_error or f"Could not determine changed files for commit {commit}. Pass --files explicitly.",
-    )
-
-
-def handle_agent_fix(repo: str | None, passthrough: list[str]) -> int:
-    parser = argparse.ArgumentParser(
-        prog="gh-address-cr agent fix",
-        description="Classify, claim, submit, and optionally publish one GitHub review-thread fix.",
-    )
-    parser.add_argument("repo")
-    parser.add_argument("pr_number")
-    parser.add_argument("item_id")
-    parser.add_argument("--agent-id", default="agent")
-    parser.add_argument("--commit", required=True)
-    parser.add_argument("--files")
-    parser.add_argument("--file", action="append", default=[])
-    parser.add_argument("--summary", required=True)
-    parser.add_argument("--why", required=True)
-    parser.add_argument("--severity", choices=["P0", "P1", "P2", "P3", "P4"])
-    parser.add_argument("--severity-note", "--severity-override-note", dest="severity_note")
-    parser.add_argument("--review-priority", choices=["high", "medium", "low"])
-    parser.add_argument("--validation", "--validation-cmd", dest="validation", action="append", default=[])
-    parser.add_argument("--publish", action="store_true")
-    parser.add_argument("--now")
-    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
-    try:
-        now_dt = None
-        if parsed.now:
-            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
-        payload = workflow.fast_fix_item(
-            parsed.repo,
-            parsed.pr_number,
-            item_id=parsed.item_id,
-            agent_id=parsed.agent_id,
-            commit_hash=parsed.commit,
-            files=_parse_agent_files(parsed.files, parsed.file),
-            validation_commands=_parse_agent_validation(parsed.validation),
-            summary=parsed.summary,
-            why=parsed.why,
-            severity=parsed.severity,
-            severity_note=parsed.severity_note,
-            review_priority=parsed.review_priority,
-            publish=parsed.publish,
-            now=now_dt,
-        )
-    except workflow.WorkflowError as exc:
-        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
-    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-    return 0
-
-
-def handle_agent_trivial_fix(repo: str | None, passthrough: list[str]) -> int:
-    parser = argparse.ArgumentParser(
-        prog="gh-address-cr agent trivial-fix",
-        description="Fast path for documentation or typo-only GitHub review-thread fixes.",
-    )
-    parser.add_argument("repo")
-    parser.add_argument("pr_number")
-    parser.add_argument("item_id")
-    parser.add_argument("--agent-id", default="agent")
-    parser.add_argument("--commit", required=True)
-    parser.add_argument("--files")
-    parser.add_argument("--file", action="append", default=[])
-    parser.add_argument("--summary", required=True)
-    parser.add_argument("--why", required=True)
-    parser.add_argument("--validation", "--validation-cmd", dest="validation", action="append", default=[])
-    parser.add_argument("--publish", action="store_true")
-    parser.add_argument("--now")
-    scope_args, scope_error = _agent_args_with_scope(repo, passthrough)
-    if scope_error is not None:
-        return _emit_scope_resolution_error(scope_error)
-    parsed = parser.parse_args(scope_args)
-    try:
-        now_dt = None
-        if parsed.now:
-            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
-        payload = workflow.trivial_fix_item(
-            parsed.repo,
-            parsed.pr_number,
-            item_id=parsed.item_id,
-            agent_id=parsed.agent_id,
-            commit_hash=parsed.commit,
-            files=_parse_agent_files(parsed.files, parsed.file),
-            validation_commands=_parse_agent_validation(parsed.validation),
-            summary=parsed.summary,
-            why=parsed.why,
-            publish=parsed.publish,
-            now=now_dt,
-        )
-    except workflow.WorkflowError as exc:
-        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
-    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-    return 0
-
-
-def handle_agent_fix_all(repo: str | None, passthrough: list[str]) -> int:
-    parser = argparse.ArgumentParser(
-        prog="gh-address-cr agent fix-all",
-        description=(
-            "Submit per-thread batch evidence for matching GitHub review threads, or use an explicit "
-            "homogeneous repeated-concern shortcut."
-        ),
-    )
-    parser.add_argument("repo")
-    parser.add_argument("pr_number")
-    parser.add_argument("--agent-id", default="agent")
-    parser.add_argument("--input", help="BatchActionResponse JSON with per-thread summary/why evidence.")
-    parser.add_argument("--commit")
-    parser.add_argument("--files")
-    parser.add_argument("--file", action="append", default=[])
-    parser.add_argument("--homogeneous-reason", help="Required rationale for the homogeneous repeated-concern shortcut.")
-    parser.add_argument("--concern-label", help="Short label for the homogeneous repeated concern.")
-    parser.add_argument("--severity", choices=["P0", "P1", "P2", "P3", "P4"])
-    parser.add_argument("--severity-note", "--severity-override-note", dest="severity_note")
-    parser.add_argument("--validation", "--validation-cmd", dest="validation", action="append", default=[])
-    parser.add_argument("--publish", action="store_true")
-    parser.add_argument("--include-stale", action="store_true")
-    parser.add_argument("--now")
-    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
-    try:
-        now_dt = None
-        if parsed.now:
-            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
-        if parsed.input:
-            conflicting_flags = []
-            if parsed.commit:
-                conflicting_flags.append("--commit")
-            if parsed.files:
-                conflicting_flags.append("--files")
-            if parsed.file:
-                conflicting_flags.append("--file")
-            if parsed.validation:
-                conflicting_flags.append("--validation")
-            if parsed.severity:
-                conflicting_flags.append("--severity")
-            if parsed.severity_note:
-                conflicting_flags.append("--severity-note")
-            if parsed.homogeneous_reason:
-                conflicting_flags.append("--homogeneous-reason")
-            if parsed.concern_label:
-                conflicting_flags.append("--concern-label")
-            if parsed.include_stale:
-                conflicting_flags.append("--include-stale")
-            if conflicting_flags:
-                raise workflow.WorkflowError(
-                    status="FAST_FIX_ALL_REJECTED",
-                    reason_code="CONFLICTING_FIX_ALL_INPUT",
-                    waiting_on="fast_fix_input",
-                    exit_code=2,
-                    message=(
-                        "agent fix-all accepts either --input or evidence flags, not both. "
-                        f"Move {', '.join(conflicting_flags)} into the BatchActionResponse."
-                    ),
-                )
-            payload = workflow.fast_fix_from_batch_input(
-                parsed.repo,
-                parsed.pr_number,
-                batch_path=parsed.input,
-                publish=parsed.publish,
-                now=now_dt,
-            )
-            sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-            return 0
-        if not parsed.commit:
-            raise workflow.WorkflowError(
-                status="FAST_FIX_ALL_REJECTED",
-                reason_code="MISSING_FIX_REPLY_COMMIT_HASH",
-                waiting_on="fast_fix_input",
-                exit_code=2,
-                message="agent fix-all requires --commit unless --input supplies a BatchActionResponse.",
-            )
-        files = _parse_agent_files(parsed.files, parsed.file)
-        if not files:
-            files = _changed_files_for_commit(parsed.commit)
-        payload = workflow.fast_fix_matching_threads(
-            parsed.repo,
-            parsed.pr_number,
-            agent_id=parsed.agent_id,
-            commit_hash=parsed.commit,
-            files=files,
-            validation_commands=_parse_agent_validation(parsed.validation),
-            include_stale=parsed.include_stale,
-            severity=parsed.severity,
-            severity_note=parsed.severity_note,
-            homogeneous_reason=parsed.homogeneous_reason,
-            concern_label=parsed.concern_label,
-            publish=parsed.publish,
-            now=now_dt,
-        )
-    except workflow.WorkflowError as exc:
-        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
-    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-    return 0
-
-
-def handle_agent_resolve_stale(repo: str | None, passthrough: list[str]) -> int:
-    parser = argparse.ArgumentParser(
-        prog="gh-address-cr agent resolve-stale",
-        description="Submit runtime-mediated evidence for stale GitHub review threads matching changed files.",
-    )
-    parser.add_argument("repo")
-    parser.add_argument("pr_number")
-    parser.add_argument("--agent-id", default="agent")
-    parser.add_argument("--commit", required=True)
-    parser.add_argument("--files")
-    parser.add_argument("--file", action="append", default=[])
-    parser.add_argument("--severity", choices=["P0", "P1", "P2", "P3", "P4"])
-    parser.add_argument("--severity-note", "--severity-override-note", dest="severity_note")
-    parser.add_argument("--validation", "--validation-cmd", dest="validation", action="append", default=[])
-    parser.add_argument("--match-files", action="store_true")
-    parser.add_argument("--publish", action="store_true")
-    parser.add_argument("--now")
-    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
-    try:
-        if not parsed.match_files:
-            raise workflow.WorkflowError(
-                status="STALE_RESOLUTION_REJECTED",
-                reason_code="MISSING_MATCH_FILES",
-                waiting_on="stale_resolution_input",
-                exit_code=2,
-                message="agent resolve-stale requires --match-files so stale synchronization stays file-scoped.",
-            )
-        now_dt = None
-        if parsed.now:
-            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
-        files = _parse_agent_files(parsed.files, parsed.file)
-        if not files:
-            files = _changed_files_for_commit(
-                parsed.commit,
-                rejected_status="STALE_RESOLUTION_REJECTED",
-                command_name="agent resolve-stale",
-            )
-        payload = workflow.fast_fix_matching_threads(
-            parsed.repo,
-            parsed.pr_number,
-            agent_id=parsed.agent_id,
-            commit_hash=parsed.commit,
-            files=files,
-            validation_commands=_parse_agent_validation(parsed.validation),
-            include_stale=True,
-            stale_only=True,
-            severity=parsed.severity,
-            severity_note=parsed.severity_note,
-            publish=parsed.publish,
-            now=now_dt,
-        )
-    except workflow.WorkflowError as exc:
-        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
-    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-    return 0
-
-
-def handle_agent_evidence(repo: str | None, passthrough: list[str]) -> int:
-    parser = argparse.ArgumentParser(prog="gh-address-cr agent evidence")
-    parser.add_argument("subcommand", choices=["add", "list"])
-    parser.add_argument("repo")
-    parser.add_argument("pr_number")
-    parser.add_argument("--name")
-    parser.add_argument("--agent-id", default="agent")
-    parser.add_argument("--commit")
-    parser.add_argument("--files")
-    parser.add_argument("--file", action="append", default=[])
-    parser.add_argument("--validation", "--validation-cmd", dest="validation", action="append", default=[])
-    parser.add_argument("--summary")
-    parser.add_argument("--why")
-    parser.add_argument("--test-command")
-    parser.add_argument("--test-result")
-    parser.add_argument("--severity", choices=["P0", "P1", "P2", "P3", "P4"])
-    parser.add_argument("--severity-note", "--severity-override-note", dest="severity_note")
-    parser.add_argument("--now")
-    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
-    try:
-        if parsed.subcommand == "list":
-            payload = workflow.list_evidence_profiles(parsed.repo, parsed.pr_number)
-        else:
-            if not parsed.name:
-                raise workflow.WorkflowError(
-                    status="EVIDENCE_PROFILE_REJECTED",
-                    reason_code="MISSING_EVIDENCE_PROFILE_NAME",
-                    waiting_on="evidence_profile",
-                    exit_code=2,
-                    message="agent evidence add requires --name.",
-                )
-            now_dt = None
-            if parsed.now:
-                now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
-            payload = workflow.record_evidence_profile(
-                parsed.repo,
-                parsed.pr_number,
-                name=parsed.name,
-                agent_id=parsed.agent_id,
-                commit_hash=parsed.commit or "",
-                files=_parse_agent_files(parsed.files, parsed.file),
-                validation_commands=_parse_agent_validation(parsed.validation),
-                summary=parsed.summary,
-                why=parsed.why,
-                test_command=parsed.test_command,
-                test_result=parsed.test_result,
-                severity=parsed.severity,
-                severity_note=parsed.severity_note,
-                now=now_dt,
-            )
-    except workflow.WorkflowError as exc:
-        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
-    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-    return 0
-
-
-def handle_agent_publish(repo: str | None, passthrough: list[str]) -> int:
-    parser = argparse.ArgumentParser(prog="gh-address-cr agent publish")
-    parser.add_argument("repo")
-    parser.add_argument("pr_number")
-    parser.add_argument("--agent-id", default="gh-address-cr-publisher")
-    parser.add_argument("--now")
-    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
-    try:
-        now_dt = None
-        if parsed.now:
-            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
-        payload = workflow.publish_github_thread_responses(
-            parsed.repo,
-            parsed.pr_number,
-            agent_id=parsed.agent_id,
-            now=now_dt,
-        )
-    except workflow.WorkflowError as exc:
-        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
-    except Exception as exc:
-        return output_generic_agent_error(parsed.repo, parsed.pr_number, "PUBLISH_ERROR", str(exc))
-    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-    return 0
-
-
-def handle_agent_leases(repo: str | None, passthrough: list[str]) -> int:
-    parser = argparse.ArgumentParser(prog="gh-address-cr agent leases")
-    parser.add_argument("repo")
-    parser.add_argument("pr_number")
-    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
-    try:
-        payload = workflow.list_leases(parsed.repo, parsed.pr_number)
-    except workflow.WorkflowError as exc:
-        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
-    except Exception as exc:
-        return output_generic_agent_error(parsed.repo, parsed.pr_number, "SESSION_ERROR", str(exc))
-    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-    return 0
-
-
-def handle_agent_reclaim(repo: str | None, passthrough: list[str]) -> int:
-    parser = argparse.ArgumentParser(prog="gh-address-cr agent reclaim")
-    parser.add_argument("repo")
-    parser.add_argument("pr_number")
-    parser.add_argument("--now")
-    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
-    now = datetime.fromisoformat(parsed.now.replace("Z", "+00:00")) if parsed.now else None
-    try:
-        payload = workflow.reclaim_leases(parsed.repo, parsed.pr_number, now=now)
-    except workflow.WorkflowError as exc:
-        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
-    except Exception as exc:
-        return output_generic_agent_error(parsed.repo, parsed.pr_number, "SESSION_ERROR", str(exc))
-    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-    return 0
-
-
-def handle_agent_orchestrate(repo: str | None, passthrough: list[str]) -> int:
-    from gh_address_cr.orchestrator import harness
-
-    return harness.handle_agent_orchestrate(repo, passthrough)
-
-
-def handle_final_gate(repo: str | None, pr_number: str | None, passthrough: list[str]) -> int:
-    if repo in {"-h", "--help"} or pr_number in {"-h", "--help"} or passthrough[:1] in (["-h"], ["--help"]):
-        print(
-            "usage: gh-address-cr final-gate <owner/repo> <pr_number> [--no-auto-clean] [--require-checks|--require-required-checks]\n\n"
-            "Host telemetry hook:\n"
-            "  GH_ADDRESS_CR_HOST_TELEMETRY_INPUT=<path> gh-address-cr final-gate <owner/repo> <pr_number>\n"
-            "  GH_ADDRESS_CR_HOST_TELEMETRY_SOURCE defaults to assistant-host.\n"
-            "  GH_ADDRESS_CR_HOST_TELEMETRY_FORMAT defaults to agent-jsonl."
-        )
-        return 0
-    parser = argparse.ArgumentParser(
-        prog="gh-address-cr final-gate",
-        formatter_class=argparse.RawDescriptionHelpFormatter,
-        epilog=(
-            "Host telemetry hook:\n"
-            "  GH_ADDRESS_CR_HOST_TELEMETRY_INPUT=<path> gh-address-cr final-gate <owner/repo> <pr_number>\n"
-            "  GH_ADDRESS_CR_HOST_TELEMETRY_SOURCE defaults to assistant-host.\n"
-            "  GH_ADDRESS_CR_HOST_TELEMETRY_FORMAT defaults to agent-jsonl.\n"
-        ),
-    )
-    output_group = parser.add_mutually_exclusive_group()
-    output_group.add_argument("--machine", action="store_true", help="Emit structured machine-readable JSON.")
-    output_group.add_argument("--human", action="store_true", help="Emit human-oriented narrative text (default).")
-    auto_group = parser.add_mutually_exclusive_group()
-    auto_group.add_argument("--auto-clean", dest="auto_clean", action="store_true")
-    auto_group.add_argument("--no-auto-clean", dest="auto_clean", action="store_false")
-    parser.set_defaults(auto_clean=True)
-    parser.add_argument("--audit-id", default="default")
-    parser.add_argument("--snapshot", default="")
-    checks_group = parser.add_mutually_exclusive_group()
-    checks_group.add_argument("--require-checks", action="store_true", help="Require all PR checks to be green.")
-    checks_group.add_argument(
-        "--require-required-checks",
-        action="store_true",
-        help="Require required PR checks to be green.",
-    )
-    parser.add_argument("repo")
-    parser.add_argument("pr_number")
-    scoped_args = _prepend_optional(repo, _prepend_optional(pr_number, passthrough))
-    scoped_args, scope_error = _maybe_prepend_implicit_scope(scoped_args)
-    if scope_error is not None:
-        return _emit_scope_resolution_error(scope_error)
-    parsed = parser.parse_args(scoped_args)
-    _ingest_host_telemetry_from_environment(parsed.repo, parsed.pr_number)
-    machine_requested = "--machine" in scoped_args and "--human" not in scoped_args
-    try:
-        result = core_gate.Gatekeeper().run(
-            parsed.repo,
-            parsed.pr_number,
-            snapshot_path=parsed.snapshot or None,
-            require_checks=parsed.require_checks,
-            require_required_checks=parsed.require_required_checks,
-        )
-    except FileNotFoundError as exc:
-        if machine_requested:
-            _emit_final_gate_machine_error(parsed.repo, parsed.pr_number, "FINAL_GATE_INPUT_MISSING", str(exc), 2)
-        else:
-            print(str(exc), file=sys.stderr)
-        return 2
-    except Exception as exc:
-        message = f"Final gate failed to evaluate: {exc}"
-        if machine_requested:
-            _emit_final_gate_machine_error(parsed.repo, parsed.pr_number, "FINAL_GATE_EVALUATION_FAILED", message, 5)
-        else:
-            print(message, file=sys.stderr)
-        return 5
-
-    summary_path, telemetry_report = _write_native_final_gate_artifacts(
-        parsed.repo, parsed.pr_number, parsed.audit_id, result
-    )
-    if result.passed and parsed.auto_clean:
-        workspace_path = session_store.workspace_dir(parsed.repo, parsed.pr_number)
-        archive_target = _archive_and_clean_workspace(parsed.repo, parsed.pr_number, parsed.audit_id)
-        if archive_target is not None:
-            summary_path = archive_target / summary_path.name
-            if telemetry_report is not None:
-                paths = core_paths.SessionPaths(parsed.repo, parsed.pr_number)
-                archived_report_path = archive_target / paths.efficiency_report_file.name
-                telemetry_report["report_artifact"] = str(archived_report_path)
-                telemetry_report = _replace_path_occurrences(
-                    telemetry_report,
-                    str(workspace_path),
-                    str(archive_target),
-                )
-                _rewrite_archived_efficiency_report_path(summary_path, telemetry_report["report_artifact"])
-                _rewrite_archived_efficiency_report_artifact(archived_report_path, telemetry_report)
-                _rewrite_archived_audit_artifacts(
-                    archive_target,
-                    original_workspace=workspace_path,
-                    summary_path=summary_path,
-                    telemetry_report=telemetry_report,
-                )
-    if parsed.machine:
-        machine_summary = result.to_machine_summary()
-        if summary_path:
-            machine_summary["artifact_path"] = str(summary_path)
-        if telemetry_report is not None:
-            machine_summary["telemetry"] = {
-                "coverage_label": telemetry_report["coverage_label"],
-                "report_artifact": telemetry_report.get("report_artifact"),
-                "total_events": telemetry_report.get("total_events"),
-                "success_rate": telemetry_report.get("success_rate"),
-                "diagnostics": telemetry_report.get("diagnostics", []),
-            }
-        sys.stdout.write(json.dumps(machine_summary, indent=2, sort_keys=True) + "\n")
-    else:
-        _emit_final_gate_result(result, summary_path=summary_path, telemetry_report=telemetry_report)
-    if not result.passed:
-        print(f"\nGate FAILED: {_final_gate_failure_message(result)}. Do not send completion summary.", file=sys.stderr)
-        return result.exit_code
-
-    return 0
-
-
-def _emit_final_gate_machine_error(repo: str, pr_number: str, reason_code: str, message: str, exit_code: int) -> None:
-    payload = {
-        "status": "BLOCKED",
-        "reason_code": reason_code,
-        "next_action": message,
-        "waiting_on": "final_gate",
-        "exit_code": exit_code,
-        "repo": repo,
-        "pr_number": str(pr_number),
-        "counts": {},
-        "failure_codes": [reason_code],
-    }
-    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-
-
-def _ingest_host_telemetry_from_environment(repo: str, pr_number: str) -> dict | None:
-    input_path = os.environ.get(HOST_TELEMETRY_INPUT_ENV)
-    if not input_path:
-        return None
-    source = os.environ.get(HOST_TELEMETRY_SOURCE_ENV) or "assistant-host"
-    fmt = os.environ.get(HOST_TELEMETRY_FORMAT_ENV) or "agent-jsonl"
-    try:
-        return _ingest_host_telemetry_input(repo, pr_number, input_path=input_path, source=source, fmt=fmt)
-    except Exception:
-        return _safe_hook_unavailable_import_summary(repo, pr_number, source=source, fmt=fmt)
-
-
-def _ingest_host_telemetry_input(
-    repo: str,
-    pr_number: str,
-    *,
-    input_path: str,
-    source: str,
-    fmt: str,
-) -> dict | None:
-    try:
-        raw = Path(input_path).read_text(encoding="utf-8")
-    except OSError:
-        return _safe_input_unavailable_import_summary(repo, pr_number, source=source, fmt=fmt)
-    return core_telemetry.import_external_telemetry(repo, pr_number, source=source, fmt=fmt, raw=raw)
-
-
-def _safe_input_unavailable_import_summary(repo: str, pr_number: str, *, source: str, fmt: str) -> dict | None:
-    try:
-        return core_telemetry.input_unavailable_import_summary(repo, pr_number, source=source, fmt=fmt)
-    except Exception:
-        return None
-
-
-def _safe_hook_unavailable_import_summary(repo: str, pr_number: str, *, source: str, fmt: str) -> dict | None:
-    try:
-        return core_telemetry.hook_unavailable_import_summary(repo, pr_number, source=source, fmt=fmt)
-    except Exception:
-        return None
-
-
-def _rewrite_archived_efficiency_report_path(summary_path: Path, report_artifact: str) -> None:
-    try:
-        lines = summary_path.read_text(encoding="utf-8").splitlines()
-    except OSError:
-        return
-    updated = [
-        f"- efficiency_report_path: {report_artifact}" if line.startswith("- efficiency_report_path:") else line
-        for line in lines
-    ]
-    try:
-        summary_path.write_text("\n".join(updated) + "\n", encoding="utf-8")
-    except OSError:
-        return
-
-
-def _rewrite_archived_efficiency_report_artifact(report_path: Path, telemetry_report: dict) -> None:
-    try:
-        current = json.loads(report_path.read_text(encoding="utf-8"))
-    except Exception:
-        current = {}
-    current.update(telemetry_report)
-    current["report_artifact"] = str(report_path)
-    try:
-        if report_path.is_dir():
-            shutil.rmtree(report_path)
-        report_path.write_text(json.dumps(current, indent=2, sort_keys=True) + "\n", encoding="utf-8")
-    except OSError:
-        return
-
-
-def _rewrite_archived_audit_artifacts(
-    archive_target: Path,
-    *,
-    original_workspace: Path,
-    summary_path: Path,
-    telemetry_report: dict,
-) -> None:
-    summary_sha256 = _read_file_sha256(summary_path)
-    for path in (archive_target / "audit.jsonl", archive_target / "trace.jsonl"):
-        try:
-            lines = path.read_text(encoding="utf-8").splitlines()
-        except OSError:
-            continue
-        updated_lines: list[str] = []
-        changed = False
-        for line in lines:
-            if not line.strip():
-                updated_lines.append(line)
-                continue
-            try:
-                entry = json.loads(line)
-            except json.JSONDecodeError:
-                updated_lines.append(line)
-                continue
-            rewritten = _replace_path_occurrences(entry, str(original_workspace), str(archive_target))
-            if isinstance(rewritten, dict):
-                if path.name == "audit.jsonl":
-                    details = rewritten.get("details")
-                    if isinstance(details, dict):
-                        details["summary_file"] = str(summary_path)
-                        details["summary_sha256"] = summary_sha256
-                        details["efficiency_report"] = dict(telemetry_report)
-                elif path.name == "trace.jsonl":
-                    rewritten["efficiency_report_path"] = telemetry_report["report_artifact"]
-            changed = changed or rewritten != entry
-            updated_lines.append(json.dumps(rewritten, sort_keys=True))
-        if changed:
-            try:
-                path.write_text("\n".join(updated_lines) + "\n", encoding="utf-8")
-            except OSError:
-                continue
-
-
-def _replace_path_occurrences(value: object, original_path: str, archived_path: str) -> object:
-    if isinstance(value, str):
-        return value.replace(original_path, archived_path)
-    if isinstance(value, list):
-        return [_replace_path_occurrences(item, original_path, archived_path) for item in value]
-    if isinstance(value, dict):
-        return {key: _replace_path_occurrences(nested, original_path, archived_path) for key, nested in value.items()}
-    return value
-
-
-def _read_file_sha256(path: Path) -> str:
-    try:
-        return hashlib.sha256(path.read_bytes()).hexdigest()
-    except OSError:
-        return "unavailable"
-
-
-def handle_telemetry_command(repo: str | None, pr_number: str | None, passthrough: list[str]) -> int:
-    if repo in {"-h", "--help"}:
-        print(
-            "usage: gh-address-cr telemetry ingest <owner/repo> <pr_number> --source <source> --format agent-jsonl --input <path>|-\n"
-            "       gh-address-cr telemetry summary <owner/repo> <pr_number> [--format json|markdown]"
-        )
-        return 0
-    if not repo:
-        print("telemetry requires a subcommand: ingest or summary", file=sys.stderr)
-        return 2
-
-    subcommand = repo
-    args = _prepend_optional(pr_number, passthrough)
-    if subcommand == "ingest":
-        parser = argparse.ArgumentParser(prog="gh-address-cr telemetry ingest")
-        parser.add_argument("repo")
-        parser.add_argument("pr_number")
-        parser.add_argument("--source", required=True)
-        parser.add_argument("--format", default="agent-jsonl")
-        parser.add_argument("--input", required=True)
-        scoped_args, scope_error = _maybe_prepend_implicit_scope(args)
-        if scope_error is not None:
-            return _emit_scope_resolution_error(scope_error)
-        parsed = parser.parse_args(scoped_args)
-        if parsed.input == "-":
-            raw = sys.stdin.read()
-        else:
-            try:
-                raw = Path(parsed.input).read_text(encoding="utf-8")
-            except OSError:
-                payload = core_telemetry.input_unavailable_import_summary(
-                    parsed.repo,
-                    parsed.pr_number,
-                    source=parsed.source,
-                    fmt=parsed.format,
-                )
-                print(json.dumps(payload, sort_keys=True))
-                return 2
-        summary = core_telemetry.import_external_telemetry(
-            parsed.repo,
-            parsed.pr_number,
-            source=parsed.source,
-            fmt=parsed.format,
-            raw=raw,
-        )
-        print(json.dumps(summary, sort_keys=True))
-        return 0 if summary["status"] in {"SUCCESS", "PARTIAL"} else 2
-
-    if subcommand == "summary":
-        parser = argparse.ArgumentParser(prog="gh-address-cr telemetry summary")
-        parser.add_argument("repo")
-        parser.add_argument("pr_number")
-        parser.add_argument("--format", choices=("json", "markdown"), default="json")
-        scoped_args, scope_error = _maybe_prepend_implicit_scope(args)
-        if scope_error is not None:
-            return _emit_scope_resolution_error(scope_error)
-        parsed = parser.parse_args(scoped_args)
-        report = core_telemetry.build_efficiency_report(parsed.repo, parsed.pr_number)
-        if _telemetry_report_has_storage_diagnostics(report):
-            payload = {
-                **report,
-                "status": "FAILED",
-                "reason_code": "TELEMETRY_REPORT_UNAVAILABLE",
-                "next_action": "FIX_TELEMETRY_STORAGE",
-            }
-            _write_telemetry_report_payload(payload)
-            print(json.dumps(payload, sort_keys=True))
-            return 2
-        if parsed.format == "markdown":
-            print(core_telemetry.efficiency_report_markdown(report), end="")
-        else:
-            print(json.dumps(report, sort_keys=True))
-        return 0
-
-    print(f"Unknown telemetry command: {subcommand}", file=sys.stderr)
-    return 2
-
-
-def _reported_telemetry_source(source: str) -> str:
-    return core_telemetry._reported_source_label(source)
-
-
-def _telemetry_report_has_storage_diagnostics(report: dict) -> bool:
-    diagnostics = report.get("diagnostics")
-    if not isinstance(diagnostics, list):
-        return False
-    return any(
-        isinstance(diagnostic, str)
-        and (
-            diagnostic.startswith("external telemetry line ")
-            or diagnostic.startswith("external telemetry unreadable:")
-            or diagnostic.startswith("external telemetry store is not a regular file:")
-            or diagnostic.startswith("telemetry import summary line ")
-            or diagnostic.startswith("telemetry import summary unreadable:")
-            or diagnostic.startswith("telemetry import summary is not a regular file:")
-            or diagnostic.startswith("efficiency report artifact unavailable:")
-        )
-        for diagnostic in diagnostics
-    )
-
-
-def _write_telemetry_report_payload(payload: dict) -> None:
-    report_artifact = payload.get("report_artifact")
-    if not isinstance(report_artifact, str) or not report_artifact:
-        return
-    path = Path(report_artifact)
-    try:
-        path.parent.mkdir(parents=True, exist_ok=True)
-        path.write_text(json.dumps(payload, indent=2, sort_keys=True) + "\n", encoding="utf-8")
-    except OSError:
-        return
-
-
-def _emit_final_gate_result(
-    result: core_gate.GateResult,
-    *,
-    summary_path: Path | None = None,
-    telemetry_report: dict | None = None,
-) -> None:
-    print("== Final Freshness Check ==")
-    print(f"Unresolved thread count: {result.counts['unresolved_remote_threads_count']}")
-    print(f"Pending review count: {result.counts['pending_current_login_review_count']}")
-    if result.check_requirement:
-        print(
-            "PR checks: "
-            f"{result.counts['pr_checks_failed_count']} failed, "
-            f"{result.counts['pr_checks_pending_count']} pending "
-            f"({result.check_requirement})"
-        )
-    print()
-    if result.passed:
-        print("Final gate PASSED")
-        print("\n== Gate Result ==")
-        print("Verified: 0 Unresolved Threads found")
-        print("Verified: 0 Pending Reviews found")
-        if result.check_requirement:
-            print("Verified: 0 Non-green PR Checks found")
-        print(f"Session blocking items: {result.counts['blocking_items_count']}")
-    else:
-        print("Final gate BLOCKED")
-        print("\n== Gate Result ==")
-        print(f"Gate FAILED: {_final_gate_failure_message(result)}")
-    print()
-    print("== Machine Gate Diagnostics ==")
-    for key in core_gate.COUNT_KEYS:
-        print(f"{key}={result.counts[key]}")
-    print(f"reason_code={result.reason_code or 'PASSED'}")
-    print(f"exit_code={result.exit_code}")
-    if telemetry_report is None:
-        telemetry_report = core_telemetry.build_efficiency_report(result.repo, result.pr_number)
-    print()
-    print("== Agent Efficiency Summary ==")
-    print(f"telemetry_coverage_label={telemetry_report['coverage_label']}")
-    print(f"telemetry_total_events={telemetry_report['total_events']}")
-    print(f"telemetry_success_rate={telemetry_report['success_rate']:.1f}")
-    print(f"telemetry_sources={_telemetry_sources_summary(telemetry_report)}")
-    print(f"telemetry_diagnostics={_telemetry_diagnostics_summary(telemetry_report)}")
-    if telemetry_report["inefficiency_flags"]:
-        print("telemetry_inefficiency_flags=" + "; ".join(telemetry_report["inefficiency_flags"]))
-    else:
-        print("telemetry_inefficiency_flags=none")
-    print(f"Efficiency report path: {telemetry_report['report_artifact']}")
-    if summary_path is not None:
-        print(f"Audit summary path: {summary_path}")
-        try:
-            summary_sha256 = hashlib.sha256(summary_path.read_bytes()).hexdigest()
-        except Exception:
-            summary_sha256 = "unavailable"
-        print(f"Audit summary sha256: {summary_sha256}")
-
-
-def _write_native_final_gate_artifacts(
-    repo: str,
-    pr_number: str,
-    audit_id: str,
-    result: core_gate.GateResult,
-) -> tuple[Path, dict]:
-    paths = core_paths.SessionPaths(repo, pr_number)
-    workspace = paths.workspace_dir
-    timestamp = datetime.now(timezone.utc).replace(microsecond=0).isoformat()
-    run_id = audit_id or "final-gate"
-    summary_path = paths.audit_summary_file
-    audit_path = paths.audit_log_file
-    trace_path = workspace / "trace.jsonl"
-    status = "ok" if result.passed else "failed"
-    summary_lines = [
-        "# Audit Summary",
-        "",
-        f"- repo: {repo}",
-        f"- pr: {pr_number}",
-        f"- run_id: {run_id}",
-        f"- final_gate_status: {status}",
-        f"- reason_code: {result.reason_code or 'PASSED'}",
-        f"- check_requirement: {result.check_requirement or 'none'}",
-    ]
-    summary_lines.extend(f"- {key}: {result.counts[key]}" for key in core_gate.COUNT_KEYS)
-    telemetry_report = core_telemetry.build_efficiency_report(repo, pr_number)
-    summary_lines.extend(
-        [
-            "",
-            "## Agent Efficiency Summary",
-            f"- telemetry_coverage_label: {telemetry_report['coverage_label']}",
-            f"- telemetry_total_events: {telemetry_report['total_events']}",
-            f"- telemetry_success_rate: {telemetry_report['success_rate']:.1f}",
-            f"- efficiency_report_path: {telemetry_report['report_artifact']}",
-            f"- telemetry_sources: {_telemetry_sources_summary(telemetry_report)}",
-            f"- telemetry_diagnostics: {_telemetry_diagnostics_summary(telemetry_report)}",
-            f"- telemetry_inefficiency_flags: {', '.join(telemetry_report['inefficiency_flags']) if telemetry_report['inefficiency_flags'] else 'none'}",
-        ]
-    )
-    if result.failure_codes:
-        summary_lines.extend(["", "## Failure Codes", *[f"- {code}" for code in result.failure_codes]])
-    summary_path.write_text("\n".join(summary_lines) + "\n", encoding="utf-8")
-    summary_sha256 = hashlib.sha256(summary_path.read_bytes()).hexdigest()
-    audit_entry = {
-        "ts": timestamp,
-        "run_id": run_id,
-        "audit_id": run_id,
-        "repo": repo,
-        "pr": str(pr_number),
-        "action": "final-gate",
-        "status": status,
-        "message": (
-            "Gate passed with zero unresolved threads"
-            if result.passed
-            else f"Gate failed; {_final_gate_failure_message(result)} remain"
-        ),
-        "details": {
-            "counts": dict(result.counts),
-            "failure_codes": list(result.failure_codes),
-            "check_requirement": result.check_requirement,
-            "summary_file": str(summary_path),
-            "summary_sha256": summary_sha256,
-            "efficiency_report": telemetry_report,
-        },
-    }
-    trace_entry = {
-        "ts": timestamp,
-        "run_id": run_id,
-        "repo": repo,
-        "pr": str(pr_number),
-        "event": "final_gate",
-        "status": status,
-        "reason_code": result.reason_code or "PASSED",
-        "counts": dict(result.counts),
-        "check_requirement": result.check_requirement,
-        "telemetry_coverage_label": telemetry_report["coverage_label"],
-        "efficiency_report_path": telemetry_report["report_artifact"],
-    }
-    for path, entry in ((audit_path, audit_entry), (trace_path, trace_entry)):
-        with path.open("a", encoding="utf-8") as handle:
-            handle.write(json.dumps(entry, sort_keys=True) + "\n")
-    return summary_path, telemetry_report
-
-
-def _telemetry_sources_summary(telemetry_report: dict) -> str:
-    sources = telemetry_report.get("sources")
-    if not isinstance(sources, list) or not sources:
-        return "none"
-    rows: list[str] = []
-    for source in sources:
-        if not isinstance(source, dict):
-            continue
-        name = source.get("source", "unknown")
-        source_type = source.get("source_type", "unknown")
-        event_count = source.get("event_count", 0)
-        coverage = source.get("coverage_status", "unknown")
-        rows.append(f"{name} ({source_type}): {event_count} events, {coverage}")
-    return "; ".join(rows) if rows else "none"
-
-
-def _telemetry_diagnostics_summary(telemetry_report: dict) -> str:
-    diagnostics = telemetry_report.get("diagnostics")
-    if not isinstance(diagnostics, list) or not diagnostics:
-        return "none"
-    return "; ".join(str(diagnostic) for diagnostic in diagnostics)
-
-
-def _final_gate_failure_message(result: core_gate.GateResult) -> str:
-    reasons: list[str] = []
-    if result.counts["unresolved_remote_threads_count"]:
-        reasons.append(f"{result.counts['unresolved_remote_threads_count']} unresolved thread(s)")
-    if result.counts["blocking_local_items_count"]:
-        reasons.append(f"{result.counts['blocking_local_items_count']} blocking item(s)")
-    if result.counts["github_threads_missing_reply_count"]:
-        reasons.append(f"{result.counts['github_threads_missing_reply_count']} GitHub thread(s) missing reply evidence")
-    if result.counts["pending_current_login_review_count"]:
-        reasons.append(f"{result.counts['pending_current_login_review_count']} pending review(s)")
-    if result.counts["missing_validation_evidence_count"]:
-        reasons.append(
-            f"{result.counts['missing_validation_evidence_count']} local item(s) missing validation evidence"
-        )
-    if result.counts["pr_checks_not_green_count"]:
-        reasons.append(f"{result.counts['pr_checks_not_green_count']} non-green PR check(s)")
-    return " and ".join(reasons) or "gate checks reported failure"
-
-
-def _archive_and_clean_workspace(repo: str, pr_number: str, audit_id: str) -> Path | None:
-    workspace = session_store.workspace_dir(repo, pr_number)
-    if not workspace.exists():
-        return None
-    archive_root = core_paths.state_dir() / "archive" / core_paths.normalize_repo(repo) / f"pr-{pr_number}"
-    archive_root.mkdir(parents=True, exist_ok=True)
-    base_name = audit_id or "final-gate"
-    archive_target = archive_root / base_name
-    suffix = 1
-    while archive_target.exists():
-        archive_target = archive_root / f"{base_name}-{suffix}"
-        suffix += 1
-    shutil.copytree(workspace, archive_target)
-    shutil.rmtree(workspace, ignore_errors=True)
-    print(f"Archived PR workspace: {archive_target}", file=sys.stderr)
-    print(f"Auto-cleaned PR workspace: {workspace}", file=sys.stderr)
-    return archive_target
-
-
-def _prepend_optional(value: str | None, args: list[str]) -> list[str]:
-    return [*([value] if value else []), *args]
-
-
-def output_workflow_error(exc: workflow.WorkflowError, *, repo: str, pr_number: str) -> int:
-    sys.stdout.write(json.dumps(exc.to_summary(repo=repo, pr_number=pr_number), indent=2, sort_keys=True) + "\n")
-    print(str(exc), file=sys.stderr)
-    return exc.exit_code
-
-
-def output_generic_agent_error(repo: str, pr_number: str, reason_code: str, message: str) -> int:
-    payload = {
-        "status": "FAILED",
-        "repo": repo,
-        "pr_number": pr_number,
-        "reason_code": reason_code,
-        "waiting_on": "session",
-        "next_action": message,
-        "exit_code": 5,
-    }
-    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-    print(message, file=sys.stderr)
-    return 5
-
-
-def _root_passthrough_args(args: argparse.Namespace) -> list[str]:
-    return [*([args.repo] if args.repo else []), *([args.pr_number] if args.pr_number else []), *args.args]
-
 
 def _emit_active_pr_payload(payload: dict, *, stderr: str | None = None) -> int:
     sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
@@ -3103,128 +1822,6 @@ def handle_active_pr_command(passthrough: list[str]) -> int:
             "next_action": f"Run `gh-address-cr address {repo} {pr_number} --lean`.",
             "exit_code": 0,
         }
-    )
-
-
-def _active_cached_sessions() -> list[tuple[str, str, Path]]:
-    try:
-        root = core_paths.state_dir()
-        if not root.exists():
-            return []
-        sessions: list[tuple[str, str, Path]] = []
-        for owner_dir in sorted(path for path in root.iterdir() if path.is_dir()):
-            if owner_dir.name == "archive" or "__" not in owner_dir.name:
-                continue
-            owner, repo_name = owner_dir.name.split("__", 1)
-            repo = f"{owner}/{repo_name}"
-            for pr_dir in sorted(path for path in owner_dir.iterdir() if path.is_dir() and path.name.startswith("pr-")):
-                pr_number = pr_dir.name.removeprefix("pr-")
-                session_path = pr_dir / "session.json"
-                if pr_number and _cached_session_is_active(session_path):
-                    sessions.append((repo, pr_number, session_path))
-        return sessions
-    except Exception:
-        return []
-
-
-def _cached_session_is_active(session_path: Path) -> bool:
-    if not session_path.is_file():
-        return False
-    try:
-        payload = json.loads(session_path.read_text(encoding="utf-8"))
-    except (OSError, json.JSONDecodeError):
-        return False
-    if not isinstance(payload, dict):
-        return False
-    return payload.get("status") == "ACTIVE"
-
-
-def _resolve_active_cached_scope() -> tuple[str, str] | dict:
-    sessions = _active_cached_sessions()
-    if not sessions:
-        return {
-            "status": "PR_SCOPE_UNRESOLVED",
-            "reason_code": "NO_ACTIVE_PR_SCOPE",
-            "waiting_on": "pr_scope",
-            "next_action": "Pass <owner/repo> <pr_number> explicitly or create exactly one cached PR session.",
-            "candidates": [],
-            "exit_code": 2,
-        }
-    if len(sessions) > 1:
-        return {
-            "status": "PR_SCOPE_UNRESOLVED",
-            "reason_code": "AMBIGUOUS_PR_SCOPE",
-            "waiting_on": "pr_scope",
-            "next_action": "Multiple cached PR sessions exist. Pass <owner/repo> <pr_number> explicitly.",
-            "candidates": [
-                {"repo": repo, "pr_number": pr_number, "session_file": str(path)}
-                for repo, pr_number, path in sessions
-            ],
-            "exit_code": 2,
-        }
-    repo, pr_number, _path = sessions[0]
-    return repo, pr_number
-
-
-def _emit_scope_resolution_error(payload: dict) -> int:
-    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
-    print(payload["next_action"], file=sys.stderr)
-    return int(payload["exit_code"])
-
-
-def _scope_positionals(args: list[str], *, value_options: set[str] | None = None) -> list[str]:
-    value_options = value_options or IMPLICIT_SCOPE_VALUE_OPTIONS
-    positional: list[str] = []
-    skip_next = False
-    for index, arg in enumerate(args):
-        if skip_next:
-            skip_next = False
-            continue
-        if arg == "--":
-            positional.extend(args[index + 1 :])
-            break
-        if arg.startswith("--"):
-            option = arg.split("=", 1)[0]
-            if option in value_options and "=" not in arg:
-                skip_next = True
-            continue
-        if arg.startswith("-"):
-            continue
-        positional.append(arg)
-    return positional
-
-
-def _maybe_prepend_implicit_scope(
-    args: list[str],
-    *,
-    value_options: set[str] | None = None,
-    allow_trailing_positionals: bool = False,
-) -> tuple[list[str], dict | None]:
-    positional = _scope_positionals(args, value_options=value_options)
-    if len(positional) >= 2:
-        return args, None
-    if positional and not allow_trailing_positionals:
-        return args, {
-            "status": "PR_SCOPE_UNRESOLVED",
-            "reason_code": "PARTIAL_PR_SCOPE",
-            "waiting_on": "pr_scope",
-            "next_action": "Pass both <owner/repo> and <pr_number>, or omit both to use the single cached PR session.",
-            "candidates": [],
-            "exit_code": 2,
-        }
-    resolved = _resolve_active_cached_scope()
-    if isinstance(resolved, dict):
-        return args, resolved
-    repo, pr_number = resolved
-    return [repo, pr_number, *args], None
-
-
-def _agent_args_with_scope(repo: str | None, passthrough: list[str]) -> tuple[list[str], dict | None]:
-    args = _prepend_optional(repo, passthrough)
-    return _maybe_prepend_implicit_scope(
-        args,
-        value_options={*IMPLICIT_SCOPE_VALUE_OPTIONS, "--summary"},
-        allow_trailing_positionals=True,
     )
