+++ src/gh_address_cr/commands/agent.py
@@ -1,0 +1,683 @@
+from __future__ import annotations
+
+import argparse
+import json
+import subprocess
+import sys
+from datetime import datetime
+
+from gh_address_cr import (
+    MAX_PARALLEL_CLAIMS,
+    PROTOCOL_VERSION,
+    SUPPORTED_PROTOCOL_VERSIONS,
+    SUPPORTED_SKILL_CONTRACT_VERSIONS,
+    __version__,
+)
+from gh_address_cr.commands.common import (
+    agent_args_with_scope as _agent_args_with_scope,
+    emit_scope_resolution_error as _emit_scope_resolution_error,
+    output_generic_agent_error,
+    output_workflow_error,
+    prepend_optional as _prepend_optional,
+)
+from gh_address_cr.core import agent_protocol, leases, publisher, workflow
+from gh_address_cr.core.errors import WorkflowError
+
+PUBLIC_COMMANDS = {
+    "active-pr",
+    "agent",
+    "address",
+    "review",
+    "threads",
+    "findings",
+    "adapter",
+    "doctor",
+    "telemetry",
+    "command-session",
+    "final-gate",
+    "review-to-findings",
+    "submit-feedback",
+    "submit-action",
+    "version",
+}
+
+
+def build_agent_manifest() -> dict:
+    return {
+        "status": "MANIFEST_READY",
+        "schema_version": PROTOCOL_VERSION,
+        "runtime_package": "gh-address-cr",
+        "runtime_version": __version__,
+        "agent_id": "gh-address-cr-runtime",
+        "protocol_versions": list(SUPPORTED_PROTOCOL_VERSIONS),
+        "supported_protocol_versions": list(SUPPORTED_PROTOCOL_VERSIONS),
+        "supported_skill_contract_versions": list(SUPPORTED_SKILL_CONTRACT_VERSIONS),
+        "roles": [
+            "coordinator",
+            "review_producer",
+            "triage",
+            "fixer",
+            "verifier",
+            "publisher",
+            "gatekeeper",
+        ],
+        "actions": [
+            "review",
+            "produce_findings",
+            "triage",
+            "classify",
+            "fix",
+            "trivial_fix",
+            "fix_all",
+            "evidence",
+            "clarify",
+            "defer",
+            "reject",
+            "verify",
+            "publish",
+            "gate",
+            "resolve_stale",
+        ],
+        "input_formats": [
+            "action_request.v1",
+            "finding.v1",
+            "github_thread.v1",
+            "evidence_profile.v1",
+            "workflow_decision.v1",
+        ],
+        "output_formats": [
+            "action_response.v1",
+            "batch_action_response.v1",
+            "evidence_record.v1",
+            "evidence_profile.v1",
+            "gate_report.v1",
+            "workflow_decision.v1",
+        ],
+        "constraints": {
+            "max_parallel_claims": MAX_PARALLEL_CLAIMS,
+        },
+        "public_commands": sorted(PUBLIC_COMMANDS),
+    }
+
+
+def handle_agent_command(args: argparse.Namespace) -> int:
+    if args.repo in {None, "-h", "--help"}:
+        sys.stdout.write(
+            "usage: gh-address-cr agent {manifest,classify,next,submit,submit-batch,fix,trivial-fix,fix-all,resolve-stale,evidence,publish,leases,reclaim,orchestrate} ...\n\n"
+            "Agent protocol utilities.\n"
+        )
+        return 0
+    if args.repo == "manifest" and not args.pr_number and not args.args:
+        sys.stdout.write(json.dumps(build_agent_manifest(), indent=2, sort_keys=True) + "\n")
+        return 0
+    if args.repo == "classify":
+        return handle_agent_classify(args.pr_number, args.args)
+    if args.repo == "next":
+        return handle_agent_next(args.pr_number, args.args)
+    if args.repo == "submit":
+        return handle_agent_submit(args.pr_number, args.args)
+    if args.repo == "submit-batch":
+        return handle_agent_submit_batch(args.pr_number, args.args)
+    if args.repo == "fix":
+        return handle_agent_fix(args.pr_number, args.args)
+    if args.repo == "trivial-fix":
+        return handle_agent_trivial_fix(args.pr_number, args.args)
+    if args.repo == "fix-all":
+        return handle_agent_fix_all(args.pr_number, args.args)
+    if args.repo == "resolve-stale":
+        return handle_agent_resolve_stale(args.pr_number, args.args)
+    if args.repo == "evidence":
+        return handle_agent_evidence(args.pr_number, args.args)
+    if args.repo == "publish":
+        return handle_agent_publish(args.pr_number, args.args)
+    if args.repo == "leases":
+        return handle_agent_leases(args.pr_number, args.args)
+    if args.repo == "reclaim":
+        return handle_agent_reclaim(args.pr_number, args.args)
+    if args.repo == "orchestrate":
+        return handle_agent_orchestrate(args.pr_number, args.args)
+    print(
+        "Unknown agent command. Supported commands: manifest, classify, next, submit, submit-batch, fix, trivial-fix, fix-all, resolve-stale, evidence, publish, leases, reclaim, orchestrate.",
+        file=sys.stderr,
+    )
+    return 2
+
+
+def handle_agent_classify(repo: str | None, passthrough: list[str]) -> int:
+    parser = argparse.ArgumentParser(prog="gh-address-cr agent classify")
+    parser.add_argument("repo")
+    parser.add_argument("pr_number")
+    parser.add_argument("item_id")
+    parser.add_argument("--classification", required=True, choices=["fix", "clarify", "defer", "reject"])
+    parser.add_argument("--agent-id", default="agent")
+    parser.add_argument("--note", required=True)
+    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
+    try:
+        payload = agent_protocol.record_classification(
+            parsed.repo,
+            parsed.pr_number,
+            item_id=parsed.item_id,
+            classification=parsed.classification,
+            agent_id=parsed.agent_id,
+            note=parsed.note,
+        )
+    except WorkflowError as exc:
+        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
+    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
+    return 0
+
+
+def handle_agent_next(repo: str | None, passthrough: list[str]) -> int:
+    parser = argparse.ArgumentParser(prog="gh-address-cr agent next")
+    parser.add_argument("repo")
+    parser.add_argument("pr_number")
+    parser.add_argument("--role", required=True)
+    parser.add_argument("--agent-id", default="agent")
+    parser.add_argument("--item-id")
+    parser.add_argument("--now")
+    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
+    try:
+        now_dt = None
+        if parsed.now:
+            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
+        payload = agent_protocol.issue_action_request(
+            parsed.repo,
+            parsed.pr_number,
+            role=parsed.role,
+            agent_id=parsed.agent_id,
+            item_id=parsed.item_id,
+            now=now_dt,
+        )
+    except WorkflowError as exc:
+        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
+    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
+    return 0
+
+
+def handle_agent_submit(repo: str | None, passthrough: list[str]) -> int:
+    parser = argparse.ArgumentParser(prog="gh-address-cr agent submit")
+    parser.add_argument("repo")
+    parser.add_argument("pr_number")
+    parser.add_argument("--input", required=True)
+    parser.add_argument("--publish", action="store_true", help="Publish accepted GitHub-thread fix evidence immediately.")
+    parser.add_argument("--now")
+    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
+    try:
+        now_dt = None
+        if parsed.now:
+            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
+        payload = agent_protocol.submit_action_response(
+            parsed.repo,
+            parsed.pr_number,
+            response_path=parsed.input,
+            now=now_dt,
+            publish=parsed.publish,
+        )
+    except WorkflowError as exc:
+        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
+    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
+    return 0
+
+
+def handle_agent_submit_batch(repo: str | None, passthrough: list[str]) -> int:
+    parser = argparse.ArgumentParser(
+        prog="gh-address-cr agent submit-batch",
+        description="Submit a BatchActionResponse with common fix evidence for multiple GitHub review threads.",
+    )
+    parser.add_argument("repo")
+    parser.add_argument("pr_number")
+    parser.add_argument("--input", required=True, help="Path to a BatchActionResponse JSON file.")
+    parser.add_argument("--now")
+    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
+    try:
+        now_dt = None
+        if parsed.now:
+            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
+        payload = agent_protocol.submit_batch_action_response(
+            parsed.repo,
+            parsed.pr_number,
+            batch_path=parsed.input,
+            now=now_dt,
+        )
+    except WorkflowError as exc:
+        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
+    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
+    return 0
+
+
+def _parse_agent_files(files: str | None, extra_files: list[str] | None = None) -> list[str]:
+    values: list[str] = []
+    if files:
+        values.extend(part.strip() for part in files.split(",") if part.strip())
+    for item in extra_files or []:
+        values.extend(part.strip() for part in item.split(",") if part.strip())
+    return values
+
+
+def _parse_agent_validation(values: list[str] | None) -> list[dict[str, str]]:
+    commands: list[dict[str, str]] = []
+    for raw in values or []:
+        command, result = _split_agent_validation_record(raw.strip())
+        if command and result:
+            commands.append({"command": command, "result": result})
+    return commands
+
+
+def _split_agent_validation_record(raw: str) -> tuple[str, str]:
+    command, separator, result = raw.rpartition("=")
+    if not separator or not _looks_like_agent_validation_result(result):
+        return raw.strip(), "passed"
+    return command.strip(), result.strip()
+
+
+def _looks_like_agent_validation_result(value: str) -> bool:
+    normalized = value.strip().lower()
+    if not normalized or any(char.isspace() for char in normalized):
+        return False
+    return normalized in {"pass", "passed", "success", "succeeded", "ok", "fail", "failed", "error", "skipped"}
+
+
+def _changed_files_for_commit(
+    commit_hash: str,
+    *,
+    rejected_status: str = "FAST_FIX_ALL_REJECTED",
+    command_name: str = "agent fix-all",
+) -> list[str]:
+    commit = commit_hash.strip()
+    if not commit:
+        return []
+    if commit.startswith("-"):
+        raise WorkflowError(
+            status=rejected_status,
+            reason_code="INVALID_COMMIT_HASH",
+            waiting_on="git_commit",
+            exit_code=2,
+            message=f"{command_name} requires a commit-ish that does not start with '-'.",
+        )
+    commands = [
+        ["git", "diff-tree", "--no-commit-id", "--name-only", "-r", commit],
+        ["git", "show", "--format=", "--name-only", commit],
+    ]
+    last_error = ""
+    for command in commands:
+        result = subprocess.run(command, text=True, capture_output=True)
+        if result.returncode != 0:
+            last_error = result.stderr.strip() or result.stdout.strip()
+            continue
+        files = [line.strip() for line in result.stdout.splitlines() if line.strip()]
+        if files:
+            return files
+    raise WorkflowError(
+        status=rejected_status,
+        reason_code="COMMIT_FILES_UNAVAILABLE",
+        waiting_on="git_commit",
+        exit_code=2,
+        message=last_error or f"Could not determine changed files for commit {commit}. Pass --files explicitly.",
+    )
+
+
+def handle_agent_fix(repo: str | None, passthrough: list[str]) -> int:
+    parser = argparse.ArgumentParser(
+        prog="gh-address-cr agent fix",
+        description="Classify, claim, submit, and optionally publish one GitHub review-thread fix.",
+    )
+    parser.add_argument("repo")
+    parser.add_argument("pr_number")
+    parser.add_argument("item_id")
+    parser.add_argument("--agent-id", default="agent")
+    parser.add_argument("--commit", required=True)
+    parser.add_argument("--files")
+    parser.add_argument("--file", action="append", default=[])
+    parser.add_argument("--summary", required=True)
+    parser.add_argument("--why", required=True)
+    parser.add_argument("--severity", choices=["P0", "P1", "P2", "P3", "P4"])
+    parser.add_argument("--severity-note", "--severity-override-note", dest="severity_note")
+    parser.add_argument("--review-priority", choices=["high", "medium", "low"])
+    parser.add_argument("--validation", "--validation-cmd", dest="validation", action="append", default=[])
+    parser.add_argument("--publish", action="store_true")
+    parser.add_argument("--now")
+    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
+    try:
+        now_dt = None
+        if parsed.now:
+            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
+        payload = workflow.fast_fix_item(
+            parsed.repo,
+            parsed.pr_number,
+            item_id=parsed.item_id,
+            agent_id=parsed.agent_id,
+            commit_hash=parsed.commit,
+            files=_parse_agent_files(parsed.files, parsed.file),
+            validation_commands=_parse_agent_validation(parsed.validation),
+            summary=parsed.summary,
+            why=parsed.why,
+            severity=parsed.severity,
+            severity_note=parsed.severity_note,
+            review_priority=parsed.review_priority,
+            publish=parsed.publish,
+            now=now_dt,
+        )
+    except WorkflowError as exc:
+        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
+    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
+    return 0
+
+
+def handle_agent_trivial_fix(repo: str | None, passthrough: list[str]) -> int:
+    parser = argparse.ArgumentParser(
+        prog="gh-address-cr agent trivial-fix",
+        description="Fast path for documentation or typo-only GitHub review-thread fixes.",
+    )
+    parser.add_argument("repo")
+    parser.add_argument("pr_number")
+    parser.add_argument("item_id")
+    parser.add_argument("--agent-id", default="agent")
+    parser.add_argument("--commit", required=True)
+    parser.add_argument("--files")
+    parser.add_argument("--file", action="append", default=[])
+    parser.add_argument("--summary", required=True)
+    parser.add_argument("--why", required=True)
+    parser.add_argument("--validation", "--validation-cmd", dest="validation", action="append", default=[])
+    parser.add_argument("--publish", action="store_true")
+    parser.add_argument("--now")
+    scope_args, scope_error = _agent_args_with_scope(repo, passthrough)
+    if scope_error is not None:
+        return _emit_scope_resolution_error(scope_error)
+    parsed = parser.parse_args(scope_args)
+    try:
+        now_dt = None
+        if parsed.now:
+            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
+        payload = workflow.trivial_fix_item(
+            parsed.repo,
+            parsed.pr_number,
+            item_id=parsed.item_id,
+            agent_id=parsed.agent_id,
+            commit_hash=parsed.commit,
+            files=_parse_agent_files(parsed.files, parsed.file),
+            validation_commands=_parse_agent_validation(parsed.validation),
+            summary=parsed.summary,
+            why=parsed.why,
+            publish=parsed.publish,
+            now=now_dt,
+        )
+    except WorkflowError as exc:
+        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
+    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
+    return 0
+
+
+def handle_agent_fix_all(repo: str | None, passthrough: list[str]) -> int:
+    parser = argparse.ArgumentParser(
+        prog="gh-address-cr agent fix-all",
+        description=(
+            "Submit per-thread batch evidence for matching GitHub review threads, or use an explicit "
+            "homogeneous repeated-concern shortcut."
+        ),
+    )
+    parser.add_argument("repo")
+    parser.add_argument("pr_number")
+    parser.add_argument("--agent-id", default="agent")
+    parser.add_argument("--input", help="BatchActionResponse JSON with per-thread summary/why evidence.")
+    parser.add_argument("--commit")
+    parser.add_argument("--files")
+    parser.add_argument("--file", action="append", default=[])
+    parser.add_argument("--homogeneous-reason", help="Required rationale for the homogeneous repeated-concern shortcut.")
+    parser.add_argument("--concern-label", help="Short label for the homogeneous repeated concern.")
+    parser.add_argument("--severity", choices=["P0", "P1", "P2", "P3", "P4"])
+    parser.add_argument("--severity-note", "--severity-override-note", dest="severity_note")
+    parser.add_argument("--validation", "--validation-cmd", dest="validation", action="append", default=[])
+    parser.add_argument("--publish", action="store_true")
+    parser.add_argument("--include-stale", action="store_true")
+    parser.add_argument("--now")
+    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
+    try:
+        now_dt = None
+        if parsed.now:
+            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
+        if parsed.input:
+            conflicting_flags = []
+            if parsed.commit:
+                conflicting_flags.append("--commit")
+            if parsed.files:
+                conflicting_flags.append("--files")
+            if parsed.file:
+                conflicting_flags.append("--file")
+            if parsed.validation:
+                conflicting_flags.append("--validation")
+            if parsed.severity:
+                conflicting_flags.append("--severity")
+            if parsed.severity_note:
+                conflicting_flags.append("--severity-note")
+            if parsed.homogeneous_reason:
+                conflicting_flags.append("--homogeneous-reason")
+            if parsed.concern_label:
+                conflicting_flags.append("--concern-label")
+            if parsed.include_stale:
+                conflicting_flags.append("--include-stale")
+            if conflicting_flags:
+                raise WorkflowError(
+                    status="FAST_FIX_ALL_REJECTED",
+                    reason_code="CONFLICTING_FIX_ALL_INPUT",
+                    waiting_on="fast_fix_input",
+                    exit_code=2,
+                    message=(
+                        "agent fix-all accepts either --input or evidence flags, not both. "
+                        f"Move {', '.join(conflicting_flags)} into the BatchActionResponse."
+                    ),
+                )
+            payload = workflow.fast_fix_from_batch_input(
+                parsed.repo,
+                parsed.pr_number,
+                batch_path=parsed.input,
+                publish=parsed.publish,
+                now=now_dt,
+            )
+            sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
+            return 0
+        if not parsed.commit:
+            raise WorkflowError(
+                status="FAST_FIX_ALL_REJECTED",
+                reason_code="MISSING_FIX_REPLY_COMMIT_HASH",
+                waiting_on="fast_fix_input",
+                exit_code=2,
+                message="agent fix-all requires --commit unless --input supplies a BatchActionResponse.",
+            )
+        files = _parse_agent_files(parsed.files, parsed.file)
+        if not files:
+            files = _changed_files_for_commit(parsed.commit)
+        payload = workflow.fast_fix_matching_threads(
+            parsed.repo,
+            parsed.pr_number,
+            agent_id=parsed.agent_id,
+            commit_hash=parsed.commit,
+            files=files,
+            validation_commands=_parse_agent_validation(parsed.validation),
+            include_stale=parsed.include_stale,
+            severity=parsed.severity,
+            severity_note=parsed.severity_note,
+            homogeneous_reason=parsed.homogeneous_reason,
+            concern_label=parsed.concern_label,
+            publish=parsed.publish,
+            now=now_dt,
+        )
+    except WorkflowError as exc:
+        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
+    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
+    return 0
+
+
+def handle_agent_resolve_stale(repo: str | None, passthrough: list[str]) -> int:
+    parser = argparse.ArgumentParser(
+        prog="gh-address-cr agent resolve-stale",
+        description="Submit runtime-mediated evidence for stale GitHub review threads matching changed files.",
+    )
+    parser.add_argument("repo")
+    parser.add_argument("pr_number")
+    parser.add_argument("--agent-id", default="agent")
+    parser.add_argument("--commit", required=True)
+    parser.add_argument("--files")
+    parser.add_argument("--file", action="append", default=[])
+    parser.add_argument("--severity", choices=["P0", "P1", "P2", "P3", "P4"])
+    parser.add_argument("--severity-note", "--severity-override-note", dest="severity_note")
+    parser.add_argument("--validation", "--validation-cmd", dest="validation", action="append", default=[])
+    parser.add_argument("--match-files", action="store_true")
+    parser.add_argument("--publish", action="store_true")
+    parser.add_argument("--now")
+    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
+    try:
+        if not parsed.match_files:
+            raise WorkflowError(
+                status="STALE_RESOLUTION_REJECTED",
+                reason_code="MISSING_MATCH_FILES",
+                waiting_on="stale_resolution_input",
+                exit_code=2,
+                message="agent resolve-stale requires --match-files so stale synchronization stays file-scoped.",
+            )
+        now_dt = None
+        if parsed.now:
+            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
+        files = _parse_agent_files(parsed.files, parsed.file)
+        if not files:
+            files = _changed_files_for_commit(
+                parsed.commit,
+                rejected_status="STALE_RESOLUTION_REJECTED",
+                command_name="agent resolve-stale",
+            )
+        payload = workflow.fast_fix_matching_threads(
+            parsed.repo,
+            parsed.pr_number,
+            agent_id=parsed.agent_id,
+            commit_hash=parsed.commit,
+            files=files,
+            validation_commands=_parse_agent_validation(parsed.validation),
+            include_stale=True,
+            stale_only=True,
+            severity=parsed.severity,
+            severity_note=parsed.severity_note,
+            publish=parsed.publish,
+            now=now_dt,
+        )
+    except WorkflowError as exc:
+        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
+    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
+    return 0
+
+
+def handle_agent_evidence(repo: str | None, passthrough: list[str]) -> int:
+    parser = argparse.ArgumentParser(prog="gh-address-cr agent evidence")
+    parser.add_argument("subcommand", choices=["add", "list"])
+    parser.add_argument("repo")
+    parser.add_argument("pr_number")
+    parser.add_argument("--name")
+    parser.add_argument("--agent-id", default="agent")
+    parser.add_argument("--commit")
+    parser.add_argument("--files")
+    parser.add_argument("--file", action="append", default=[])
+    parser.add_argument("--validation", "--validation-cmd", dest="validation", action="append", default=[])
+    parser.add_argument("--summary")
+    parser.add_argument("--why")
+    parser.add_argument("--test-command")
+    parser.add_argument("--test-result")
+    parser.add_argument("--severity", choices=["P0", "P1", "P2", "P3", "P4"])
+    parser.add_argument("--severity-note", "--severity-override-note", dest="severity_note")
+    parser.add_argument("--now")
+    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
+    try:
+        if parsed.subcommand == "list":
+            payload = workflow.list_evidence_profiles(parsed.repo, parsed.pr_number)
+        else:
+            if not parsed.name:
+                raise WorkflowError(
+                    status="EVIDENCE_PROFILE_REJECTED",
+                    reason_code="MISSING_EVIDENCE_PROFILE_NAME",
+                    waiting_on="evidence_profile",
+                    exit_code=2,
+                    message="agent evidence add requires --name.",
+                )
+            now_dt = None
+            if parsed.now:
+                now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
+            payload = workflow.record_evidence_profile(
+                parsed.repo,
+                parsed.pr_number,
+                name=parsed.name,
+                agent_id=parsed.agent_id,
+                commit_hash=parsed.commit or "",
+                files=_parse_agent_files(parsed.files, parsed.file),
+                validation_commands=_parse_agent_validation(parsed.validation),
+                summary=parsed.summary,
+                why=parsed.why,
+                test_command=parsed.test_command,
+                test_result=parsed.test_result,
+                severity=parsed.severity,
+                severity_note=parsed.severity_note,
+                now=now_dt,
+            )
+    except WorkflowError as exc:
+        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
+    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
+    return 0
+
+
+def handle_agent_publish(repo: str | None, passthrough: list[str]) -> int:
+    parser = argparse.ArgumentParser(prog="gh-address-cr agent publish")
+    parser.add_argument("repo")
+    parser.add_argument("pr_number")
+    parser.add_argument("--agent-id", default="gh-address-cr-publisher")
+    parser.add_argument("--now")
+    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
+    try:
+        now_dt = None
+        if parsed.now:
+            now_dt = datetime.fromisoformat(parsed.now.replace("Z", "+00:00"))
+        payload = publisher.publish_github_thread_responses(
+            parsed.repo,
+            parsed.pr_number,
+            agent_id=parsed.agent_id,
+            now=now_dt,
+        )
+    except WorkflowError as exc:
+        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
+    except Exception as exc:
+        return output_generic_agent_error(parsed.repo, parsed.pr_number, "PUBLISH_ERROR", str(exc))
+    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
+    return 0
+
+
+def handle_agent_leases(repo: str | None, passthrough: list[str]) -> int:
+    parser = argparse.ArgumentParser(prog="gh-address-cr agent leases")
+    parser.add_argument("repo")
+    parser.add_argument("pr_number")
+    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
+    try:
+        payload = leases.list_leases(parsed.repo, parsed.pr_number)
+    except WorkflowError as exc:
+        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
+    except Exception as exc:
+        return output_generic_agent_error(parsed.repo, parsed.pr_number, "SESSION_ERROR", str(exc))
+    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
+    return 0
+
+
+def handle_agent_reclaim(repo: str | None, passthrough: list[str]) -> int:
+    parser = argparse.ArgumentParser(prog="gh-address-cr agent reclaim")
+    parser.add_argument("repo")
+    parser.add_argument("pr_number")
+    parser.add_argument("--now")
+    parsed = parser.parse_args(_prepend_optional(repo, passthrough))
+    now = datetime.fromisoformat(parsed.now.replace("Z", "+00:00")) if parsed.now else None
+    try:
+        payload = leases.reclaim_leases(parsed.repo, parsed.pr_number, now=now)
+    except WorkflowError as exc:
+        return output_workflow_error(exc, repo=parsed.repo, pr_number=parsed.pr_number)
+    except Exception as exc:
+        return output_generic_agent_error(parsed.repo, parsed.pr_number, "SESSION_ERROR", str(exc))
+    sys.stdout.write(json.dumps(payload, indent=2, sort_keys=True) + "\n")
+    return 0
+
+
+def handle_agent_orchestrate(repo: str | None, passthrough: list[str]) -> int:
+    from gh_address_cr.orchestrator import harness
+
+    return harness.handle_agent_orchestrate(repo, passthrough)
