+++ tests/test_native_workflow.py
@@ -6,6 +6,8 @@ from datetime import datetime, timedelta, timezone
 from pathlib import Path
 from unittest.mock import patch
 
+from gh_address_cr.core import agent_protocol, publisher
+
 
 def open_item(item_id="local:1"):
     return {
@@ -34,7 +36,7 @@ class NativeWorkflowTests(unittest.TestCase):
         return manager
 
     def test_record_classification_unblocks_fixer_action_request(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import agent_protocol
 
         repo = "owner/repo"
         pr_number = "123"
@@ -42,7 +44,7 @@ class NativeWorkflowTests(unittest.TestCase):
             with patch.dict(os.environ, {"GH_ADDRESS_CR_STATE_DIR": tmp}, clear=False):
                 manager = self.write_session(repo, pr_number, open_item())
 
-                classified = workflow.record_classification(
+                classified = agent_protocol.record_classification(
                     repo,
                     pr_number,
                     item_id="local:1",
@@ -50,7 +52,7 @@ class NativeWorkflowTests(unittest.TestCase):
                     agent_id="triage-1",
                     note="Real defect.",
                 )
-                requested = workflow.issue_action_request(repo, pr_number, role="fixer", agent_id="fixer-1")
+                requested = agent_protocol.issue_action_request(repo, pr_number, role="fixer", agent_id="fixer-1")
 
                 session = manager.load()
                 evidence_rows = [
@@ -63,7 +65,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.assertEqual(evidence_rows[0]["agent_id"], "triage-1")
 
     def test_record_classification_releases_active_triage_lease_for_fixer(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import agent_protocol
 
         repo = "owner/repo"
         pr_number = "123"
@@ -71,8 +73,8 @@ class NativeWorkflowTests(unittest.TestCase):
             with patch.dict(os.environ, {"GH_ADDRESS_CR_STATE_DIR": tmp}, clear=False):
                 manager = self.write_session(repo, pr_number, open_item())
 
-                triage = workflow.issue_action_request(repo, pr_number, role="triage", agent_id="triage-1")
-                classified = workflow.record_classification(
+                triage = agent_protocol.issue_action_request(repo, pr_number, role="triage", agent_id="triage-1")
+                classified = agent_protocol.record_classification(
                     repo,
                     pr_number,
                     item_id="local:1",
@@ -80,7 +82,7 @@ class NativeWorkflowTests(unittest.TestCase):
                     agent_id="triage-1",
                     note="Real defect.",
                 )
-                fixer = workflow.issue_action_request(repo, pr_number, role="fixer", agent_id="fixer-1")
+                fixer = agent_protocol.issue_action_request(repo, pr_number, role="fixer", agent_id="fixer-1")
 
                 session = manager.load()
                 triage_lease = session["leases"][triage["lease_id"]]
@@ -94,7 +96,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.assertEqual(session["items"]["local:1"]["active_lease_id"], fixer["lease_id"])
 
     def test_record_classification_rejects_unknown_item_without_mutation(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import agent_protocol
 
         repo = "owner/repo"
         pr_number = "123"
@@ -102,8 +104,8 @@ class NativeWorkflowTests(unittest.TestCase):
             with patch.dict(os.environ, {"GH_ADDRESS_CR_STATE_DIR": tmp}, clear=False):
                 manager = self.write_session(repo, pr_number, open_item())
 
-                with self.assertRaises(workflow.WorkflowError) as context:
-                    workflow.record_classification(
+                with self.assertRaises(agent_protocol.WorkflowError) as context:
+                    agent_protocol.record_classification(
                         repo,
                         pr_number,
                         item_id="missing",
@@ -117,7 +119,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.assertNotIn("classification_evidence", session["items"]["local:1"])
 
     def test_record_classification_rejects_unsupported_classification(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import agent_protocol
 
         repo = "owner/repo"
         pr_number = "123"
@@ -125,8 +127,8 @@ class NativeWorkflowTests(unittest.TestCase):
             with patch.dict(os.environ, {"GH_ADDRESS_CR_STATE_DIR": tmp}, clear=False):
                 self.write_session(repo, pr_number, open_item())
 
-                with self.assertRaises(workflow.WorkflowError) as context:
-                    workflow.record_classification(
+                with self.assertRaises(agent_protocol.WorkflowError) as context:
+                    agent_protocol.record_classification(
                         repo,
                         pr_number,
                         item_id="local:1",
@@ -138,7 +140,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.assertEqual(context.exception.reason_code, "UNSUPPORTED_CLASSIFICATION")
 
     def test_publish_github_thread_response_posts_reply_resolves_and_closes_item(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import publisher
 
         class FakeGitHubClient:
             def __init__(self):
@@ -178,7 +180,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 manager = self.write_session(repo, pr_number, item)
                 client = FakeGitHubClient()
 
-                result = workflow.publish_github_thread_responses(repo, pr_number, github_client=client)
+                result = publisher.publish_github_thread_responses(repo, pr_number, github_client=client)
 
                 session = manager.load()
                 updated = session["items"]["github-thread:THREAD_1"]
@@ -213,7 +215,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.assertIn("response_published", event_types)
 
     def test_publish_mixed_review_threads_posts_distinct_targeted_replies(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import publisher
 
         class FakeGitHubClient:
             def __init__(self):
@@ -280,7 +282,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 manager.save(session)
                 client = FakeGitHubClient()
 
-                result = workflow.publish_github_thread_responses(repo, pr_number, github_client=client)
+                result = publisher.publish_github_thread_responses(repo, pr_number, github_client=client)
 
                 self.assertEqual(result["status"], "PUBLISH_COMPLETE")
                 self.assertEqual(result["published_count"], 2)
@@ -293,7 +295,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.assertIn("The logging path now omits the sensitive token mentioned in this thread.", second_body)
 
     def test_submit_action_response_with_publish_posts_and_resolves_thread(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import agent_protocol
 
         class FakeGitHubClient:
             def __init__(self):
@@ -326,7 +328,7 @@ class NativeWorkflowTests(unittest.TestCase):
         with tempfile.TemporaryDirectory() as tmp:
             with patch.dict(os.environ, {"GH_ADDRESS_CR_STATE_DIR": tmp}, clear=False):
                 manager = self.write_session(repo, pr_number, item)
-                request_info = workflow.issue_action_request(repo, pr_number, role="fixer", agent_id="codex-1")
+                request_info = agent_protocol.issue_action_request(repo, pr_number, role="fixer", agent_id="codex-1")
                 request = json.loads(Path(request_info["request_path"]).read_text(encoding="utf-8"))
                 response_path = Path(tmp) / "action-response.json"
                 response_path.write_text(
@@ -354,7 +356,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 )
                 client = FakeGitHubClient()
 
-                result = workflow.submit_action_response(
+                result = agent_protocol.submit_action_response(
                     repo,
                     pr_number,
                     response_path=response_path,
@@ -371,7 +373,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.assertEqual(session["items"]["github-thread:THREAD_1"]["reply_evidence"]["author_login"], "agent-login")
 
     def test_publish_with_no_ready_items_does_not_require_viewer_login(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import publisher
 
         class FakeGitHubClient:
             def viewer_login(self):
@@ -383,12 +385,12 @@ class NativeWorkflowTests(unittest.TestCase):
             with patch.dict(os.environ, {"GH_ADDRESS_CR_STATE_DIR": tmp}, clear=False):
                 self.write_session(repo, pr_number, open_item())
 
-                result = workflow.publish_github_thread_responses(repo, pr_number, github_client=FakeGitHubClient())
+                result = publisher.publish_github_thread_responses(repo, pr_number, github_client=FakeGitHubClient())
 
                 self.assertEqual(result["status"], "NO_PUBLISH_READY_ITEMS")
 
     def test_publish_ready_thread_survives_remote_refresh_before_publish(self):
-        from gh_address_cr.core import gate, workflow
+        from gh_address_cr.core import gate
 
         class FakeGitHubClient:
             def __init__(self):
@@ -445,7 +447,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 manager.save(refreshed)
                 client = FakeGitHubClient()
 
-                result = workflow.publish_github_thread_responses(repo, pr_number, github_client=client)
+                result = publisher.publish_github_thread_responses(repo, pr_number, github_client=client)
 
                 session = manager.load()
                 self.assertEqual(result["status"], "PUBLISH_COMPLETE")
@@ -454,7 +456,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.assertEqual(session["items"]["github-thread:THREAD_1"]["state"], "closed")
 
     def test_remote_refresh_recovers_publish_ready_from_accepted_evidence(self):
-        from gh_address_cr.core import gate, workflow
+        from gh_address_cr.core import gate
 
         class FakeGitHubClient:
             def __init__(self):
@@ -513,7 +515,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.assertEqual(refreshed["items"]["github-thread:THREAD_1"]["state"], "publish_ready")
 
                 client = FakeGitHubClient()
-                result = workflow.publish_github_thread_responses(repo, pr_number, github_client=client)
+                result = publisher.publish_github_thread_responses(repo, pr_number, github_client=client)
 
                 session = manager.load()
                 self.assertEqual(result["status"], "PUBLISH_COMPLETE")
@@ -522,7 +524,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.assertEqual(session["items"]["github-thread:THREAD_1"]["state"], "closed")
 
     def test_publish_github_thread_fix_uses_documented_reply_template(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import publisher
 
         class FakeGitHubClient:
             def __init__(self):
@@ -587,14 +589,14 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.write_session(repo, pr_number, item)
                 client = FakeGitHubClient()
 
-                workflow.publish_github_thread_responses(repo, pr_number, github_client=client)
+                publisher.publish_github_thread_responses(repo, pr_number, github_client=client)
 
                 self.assertEqual(client.replies[0], expected)
                 self.assertNotIn("Severity:", client.replies[0])
                 self.assertNotIn("Reviewer priority:", client.replies[0])
 
     def test_publish_github_thread_fix_without_severity_does_not_default_to_p2(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import publisher
 
         class FakeGitHubClient:
             def __init__(self):
@@ -636,14 +638,14 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.write_session(repo, pr_number, item)
                 client = FakeGitHubClient()
 
-                workflow.publish_github_thread_responses(repo, pr_number, github_client=client)
+                publisher.publish_github_thread_responses(repo, pr_number, github_client=client)
 
                 self.assertNotIn("Severity:", client.replies[0])
                 self.assertNotIn("Review signal:", client.replies[0])
                 self.assertNotIn("Medium-severity path", client.replies[0])
 
     def test_publish_github_thread_fix_surfaces_raw_reviewer_priority(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import publisher
 
         class FakeGitHubClient:
             def __init__(self):
@@ -691,7 +693,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.write_session(repo, pr_number, item)
                 client = FakeGitHubClient()
 
-                workflow.publish_github_thread_responses(repo, pr_number, github_client=client)
+                publisher.publish_github_thread_responses(repo, pr_number, github_client=client)
 
                 self.assertNotIn("Severity:", client.replies[0])
                 self.assertNotIn("Reviewer priority:", client.replies[0])
@@ -700,7 +702,7 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.assertIn("Low-priority reviewer signal", client.replies[0])
 
     def test_publish_github_thread_fix_with_legacy_unbacked_severity_does_not_default_to_p2(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import publisher
 
         class FakeGitHubClient:
             def __init__(self):
@@ -743,13 +745,13 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.write_session(repo, pr_number, item)
                 client = FakeGitHubClient()
 
-                workflow.publish_github_thread_responses(repo, pr_number, github_client=client)
+                publisher.publish_github_thread_responses(repo, pr_number, github_client=client)
 
                 self.assertNotIn("Severity:", client.replies[0])
                 self.assertNotIn("Medium-severity path", client.replies[0])
 
     def test_publish_github_thread_fix_uses_severity_specific_template_lines(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import publisher
 
         class FakeGitHubClient:
             def __init__(self):
@@ -805,7 +807,7 @@ class NativeWorkflowTests(unittest.TestCase):
                         self.write_session(repo, pr_number, item)
                         client = FakeGitHubClient()
 
-                        workflow.publish_github_thread_responses(repo, pr_number, github_client=client)
+                        publisher.publish_github_thread_responses(repo, pr_number, github_client=client)
 
                         self.assertIn(f"Review signal: `{severity}`", client.replies[0])
                         self.assertNotIn("Severity:", client.replies[0])
@@ -813,7 +815,7 @@ class NativeWorkflowTests(unittest.TestCase):
                         self.assertIn(expected_line, client.replies[0])
 
     def test_publish_github_thread_fix_ignores_reply_markdown_when_fix_reply_exists(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import publisher
 
         class FakeGitHubClient:
             def __init__(self):
@@ -856,13 +858,13 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.write_session(repo, pr_number, item)
                 client = FakeGitHubClient()
 
-                workflow.publish_github_thread_responses(repo, pr_number, github_client=client)
+                publisher.publish_github_thread_responses(repo, pr_number, github_client=client)
 
                 self.assertIn("Fixed in `abc123`.", client.replies[0])
                 self.assertNotIn("Legacy handwritten reply", client.replies[0])
 
     def test_publish_github_thread_defer_uses_documented_reply_template(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import publisher
 
         class FakeGitHubClient:
             def __init__(self):
@@ -910,12 +912,12 @@ class NativeWorkflowTests(unittest.TestCase):
                 self.write_session(repo, pr_number, item)
                 client = FakeGitHubClient()
 
-                workflow.publish_github_thread_responses(repo, pr_number, github_client=client)
+                publisher.publish_github_thread_responses(repo, pr_number, github_client=client)
 
                 self.assertEqual(client.replies[0], expected)
 
     def test_publish_github_thread_response_fails_before_side_effect_without_reply_body(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import publisher
 
         class FakeGitHubClient:
             def post_reply(self, repo, pr_number, thread_id, body):
@@ -944,8 +946,8 @@ class NativeWorkflowTests(unittest.TestCase):
             with patch.dict(os.environ, {"GH_ADDRESS_CR_STATE_DIR": tmp}, clear=False):
                 manager = self.write_session(repo, pr_number, item)
 
-                with self.assertRaises(workflow.WorkflowError) as context:
-                    workflow.publish_github_thread_responses(
+                with self.assertRaises(agent_protocol.WorkflowError) as context:
+                    publisher.publish_github_thread_responses(
                         repo,
                         pr_number,
                         github_client=FakeGitHubClient(),
@@ -985,7 +987,7 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
         return manager
 
     def test_stale_github_thread_is_claimable_by_triage_role(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import agent_protocol
 
         repo = "owner/repo"
         pr_number = "500"
@@ -993,13 +995,13 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
             with patch.dict(os.environ, {"GH_ADDRESS_CR_STATE_DIR": tmp}, clear=False):
                 self.write_session(repo, pr_number, stale_github_thread_item())
 
-                result = workflow.issue_action_request(repo, pr_number, role="triage", agent_id="triage-1")
+                result = agent_protocol.issue_action_request(repo, pr_number, role="triage", agent_id="triage-1")
 
                 self.assertEqual(result["status"], "ACTION_REQUESTED")
                 self.assertEqual(result["item_id"], "github-thread:THREAD_STALE")
 
     def test_stale_github_thread_with_classification_is_claimable_by_fixer_role(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import agent_protocol
 
         repo = "owner/repo"
         pr_number = "501"
@@ -1014,13 +1016,13 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
             with patch.dict(os.environ, {"GH_ADDRESS_CR_STATE_DIR": tmp}, clear=False):
                 self.write_session(repo, pr_number, item)
 
-                result = workflow.issue_action_request(repo, pr_number, role="fixer", agent_id="fixer-1")
+                result = agent_protocol.issue_action_request(repo, pr_number, role="fixer", agent_id="fixer-1")
 
                 self.assertEqual(result["status"], "ACTION_REQUESTED")
                 self.assertEqual(result["item_id"], "github-thread:THREAD_STALE")
 
     def test_stale_thread_not_claimed_by_fixer_without_classification(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import agent_protocol
 
         repo = "owner/repo"
         pr_number = "502"
@@ -1028,13 +1030,13 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
             with patch.dict(os.environ, {"GH_ADDRESS_CR_STATE_DIR": tmp}, clear=False):
                 self.write_session(repo, pr_number, stale_github_thread_item())
 
-                with self.assertRaises(workflow.WorkflowError) as context:
-                    workflow.issue_action_request(repo, pr_number, role="fixer", agent_id="fixer-1")
+                with self.assertRaises(agent_protocol.WorkflowError) as context:
+                    agent_protocol.issue_action_request(repo, pr_number, role="fixer", agent_id="fixer-1")
 
                 self.assertEqual(context.exception.reason_code, "MISSING_CLASSIFICATION")
 
     def test_stale_thread_classification_keeps_stale_status_until_fixer_claim(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import agent_protocol
 
         repo = "owner/repo"
         pr_number = "503"
@@ -1042,8 +1044,8 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
             with patch.dict(os.environ, {"GH_ADDRESS_CR_STATE_DIR": tmp}, clear=False):
                 manager = self.write_session(repo, pr_number, stale_github_thread_item())
 
-                triage = workflow.issue_action_request(repo, pr_number, role="triage", agent_id="triage-1")
-                workflow.record_classification(
+                triage = agent_protocol.issue_action_request(repo, pr_number, role="triage", agent_id="triage-1")
+                agent_protocol.record_classification(
                     repo,
                     pr_number,
                     item_id="github-thread:THREAD_STALE",
@@ -1060,7 +1062,7 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
                 self.assertNotIn("active_lease_id", item)
 
     def test_reclaim_expired_stale_thread_lease_restores_stale_state(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import agent_protocol, leases
 
         repo = "owner/repo"
         pr_number = "504"
@@ -1076,8 +1078,8 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
             with patch.dict(os.environ, {"GH_ADDRESS_CR_STATE_DIR": tmp}, clear=False):
                 manager = self.write_session(repo, pr_number, item)
 
-                workflow.issue_action_request(repo, pr_number, role="fixer", agent_id="fixer-1", now=now)
-                reclaimed = workflow.reclaim_leases(repo, pr_number, now=now + timedelta(hours=2))
+                agent_protocol.issue_action_request(repo, pr_number, role="fixer", agent_id="fixer-1", now=now)
+                reclaimed = leases.reclaim_leases(repo, pr_number, now=now + timedelta(hours=2))
 
                 session = manager.load()
                 item = session["items"]["github-thread:THREAD_STALE"]
@@ -1087,7 +1089,7 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
                 self.assertNotIn("active_lease_id", item)
 
     def test_stale_thread_classify_submit_publish_final_gate_path(self):
-        from gh_address_cr.core import gate, workflow
+        from gh_address_cr.core import gate
 
         class FakeGitHubClient:
             def __init__(self):
@@ -1107,7 +1109,7 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
         with tempfile.TemporaryDirectory() as tmp:
             with patch.dict(os.environ, {"GH_ADDRESS_CR_STATE_DIR": tmp}, clear=False):
                 manager = self.write_session(repo, pr_number, stale_github_thread_item())
-                workflow.record_classification(
+                agent_protocol.record_classification(
                     repo,
                     pr_number,
                     item_id="github-thread:THREAD_STALE",
@@ -1115,7 +1117,7 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
                     agent_id="triage-1",
                     note="Real defect, needs null guard.",
                 )
-                requested = workflow.issue_action_request(repo, pr_number, role="fixer", agent_id="fixer-1")
+                requested = agent_protocol.issue_action_request(repo, pr_number, role="fixer", agent_id="fixer-1")
                 request = json.loads(Path(requested["request_path"]).read_text(encoding="utf-8"))
                 response_path = Path(tmp) / "stale-action-response.json"
                 response_path.write_text(
@@ -1145,8 +1147,8 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
                     encoding="utf-8",
                 )
 
-                accepted = workflow.submit_action_response(repo, pr_number, response_path=response_path)
-                published = workflow.publish_github_thread_responses(
+                accepted = agent_protocol.submit_action_response(repo, pr_number, response_path=response_path)
+                published = publisher.publish_github_thread_responses(
                     repo,
                     pr_number,
                     agent_id="agent-login",
@@ -1166,7 +1168,7 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
                 self.assertEqual(result.counts["github_threads_missing_reply_count"], 0)
 
     def test_stale_thread_classify_then_fixer_claim(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import agent_protocol
 
         repo = "owner/repo"
         pr_number = "506"
@@ -1174,8 +1176,8 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
             with patch.dict(os.environ, {"GH_ADDRESS_CR_STATE_DIR": tmp}, clear=False):
                 manager = self.write_session(repo, pr_number, stale_github_thread_item())
 
-                workflow.issue_action_request(repo, pr_number, role="triage", agent_id="triage-1")
-                classified = workflow.record_classification(
+                agent_protocol.issue_action_request(repo, pr_number, role="triage", agent_id="triage-1")
+                classified = agent_protocol.record_classification(
                     repo,
                     pr_number,
                     item_id="github-thread:THREAD_STALE",
@@ -1183,7 +1185,7 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
                     agent_id="triage-1",
                     note="Real defect, needs null guard.",
                 )
-                fixer = workflow.issue_action_request(repo, pr_number, role="fixer", agent_id="fixer-1")
+                fixer = agent_protocol.issue_action_request(repo, pr_number, role="fixer", agent_id="fixer-1")
 
                 session = manager.load()
                 self.assertEqual(classified["status"], "CLASSIFICATION_RECORDED")
@@ -1193,7 +1195,7 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
                 self.assertEqual(item["classification_evidence"]["classification"], "fix")
 
     def test_publish_github_thread_responses_does_not_include_efficiency_summary(self):
-        from gh_address_cr.core import workflow
+        from gh_address_cr.core import publisher
         from gh_address_cr.core.telemetry import SessionTelemetry
 
         class FakeGitHubClient:
@@ -1238,7 +1240,7 @@ class StaleThreadClaimabilityTests(unittest.TestCase):
                 tracker.record("npm install", 100.0, 105.0, 0)
                 SessionTelemetry.reset()
 
-                workflow.publish_github_thread_responses(repo, pr_number, github_client=client)
+                publisher.publish_github_thread_responses(repo, pr_number, github_client=client)
 
                 self.assertEqual(len(client.replies), 1)
                 self.assertNotIn("Agent Efficiency Summary", client.replies[0])
